% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{anno_track}
\alias{anno_track}
\title{Format Annotation Tracks}
\usage{
anno_track(dat, features, var_type)
}
\arguments{
\item{dat}{Omic data matrix or matrix-like object with rows corresponding to
probes and columns to samples.}

\item{features}{Vector of length equal to sample size, or several such
vectors organized into a list or data frame.}

\item{var_type}{Character string specifying whether features are categorical
or continuous.}
}
\description{
This utility function formats categorical or continuous features for
plotting atop heatmaps.
}
