% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{colorize}
\alias{colorize}
\title{Assign Colors}
\usage{
colorize(pal, n, var_type)
}
\arguments{
\item{pal}{Color palette.}

\item{n}{Number of colors.}

\item{var_type}{Character string specifying whether features are categorical
or continuous.}
}
\description{
This utility function applies a user selected color palette to a ggplot
image.
}
