% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dist_mat}
\alias{dist_mat}
\title{Create Distance Matrix}
\usage{
dist_mat(dat, top, filter_method, dist)
}
\arguments{
\item{dat}{Omic data matrix or matrix-like object with rows corresponding to
probes and columns to samples.}

\item{top}{Optional number (if > 1) or proportion (if < 1) of top probes to
be used for distance calculations.}

\item{dist}{Distance measure to be used. Currently supports \code{
"euclidean", "pearson", "MI",} or \code{"KLD"}.}
}
\description{
This utility function calculates distance based on a user defined measure
and optionally filters probes by leading log fold change.
}
\references{
Ritchie, M.E., Phipson, B., Wu, D., Hu, Y., Law, C.W., Shi, W., & Smyth, G.K.
(2015).
\href{https://www.ncbi.nlm.nih.gov/pubmed/25605792}{limma powers differential
expression analyses for RNA-sequencing and microarray studies}. \emph{Nucleic
Acids Res.}, emph{43}(7): e47.
}
