% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{embed}
\alias{embed}
\title{Embed Manifold}
\usage{
embed(
  df,
  group,
  covar,
  group_cols,
  covar_cols,
  feature_names,
  label,
  size,
  alpha,
  title,
  xlab,
  ylab,
  legend,
  hover,
  D3
)
}
\arguments{
\item{df}{Data frame formatted for \code{ggplot}ing.}

\item{group}{Optional factor variable formatted by \code{format_features}.}

\item{covar}{Optional continuous variable formatted by \code{format_features}.}

\item{group_cols}{Palette for \code{group} formatted by \code{colorize}.}

\item{covar_cols}{Palette for \code{covar} formatted by \code{colorize}.}

\item{feature_names}{Feature names.}

\item{label}{Label data points by sample name?}

\item{title}{Plot title.}

\item{xlab}{X-axis label.}

\item{ylab}{Y-axis label.}

\item{legend}{Legend location.}

\item{hover}{Show sample name by hovering mouse over data point?}

\item{D3}{Render plot in three dimensions?}
}
\description{
This utility function formats the output for \code{bioplotr}'s projection
functions (\code{plot_pca}, \code{plot_mds}, and \code{plot_tsne}).
}
