% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_binom}
\alias{format_binom}
\title{Format Binomial Things}
\usage{
format_binom(vec, vec_type, n)
}
\arguments{
\item{vec}{Vector of observed outcomes, or one or several vectors of
predicted values.}

\item{vec_type}{Character string specifying whether vector is \code{"obs"}
or \code{"pred"}.}

\item{n}{Number of events for which predictions are required.}
}
\description{
This utility function formats vectors of observed and predicted values for
binomially distributed events.
}
