% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_features}
\alias{format_features}
\title{Format Features}
\usage{
format_features(dat, features, var_type)
}
\arguments{
\item{dat}{Omic data matrix or matrix-like object with rows corresponding to
probes and columns to samples.}

\item{features}{Vector of length equal to sample size, or several such
vectors organized into a list or data frame.}

\item{var_type}{Character string specifying whether features are \code{
"Categorical"} or \code{"Continuous"}.}
}
\description{
This utility function formats categorical and continuous features for plots.
}
