% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.color}
\alias{is.color}
\title{Test for Valid Color Representation}
\usage{
is.color(chr)
}
\arguments{
\item{chr}{A character vector.}
}
\description{
This utility function checks whether each string in a character vector
denotes a valid color in R.
}
