% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{kpca_fn}
\alias{kpca_fn}
\title{Compute kernel principal components}
\usage{
kpca_fn(kernel, kpar)
}
\arguments{
\item{kernel}{The kernel generating function. Options include 
\code{"rbfdot"}, \code{"polydot"}, \code{"tanhdot"}, \code{"vanilladot"}, 
\code{"laplacedot"}, \code{"besseldot"}, \code{"anovadot"}, and 
\code{"splinedot"}.}

\item{kpar}{A named list of arguments setting parameters for the kernel
function.}
}
\description{
This utility function computes the kernel matrix and performs 
eigendecomposition.
}
