% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcpm.R
\name{lcpm}
\alias{lcpm}
\title{Transform Counts}
\usage{
lcpm(mat)
}
\description{
This convenient wrapper function converts raw counts (or count-like data) to the
log2-counts per million scale, following RLE normalization.
}
