% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{locate_legend}
\alias{locate_legend}
\title{Locate Legend}
\usage{
locate_legend(p, loc)
}
\arguments{
\item{p}{A \code{ggplot2} object.}

\item{loc}{String specifying legend location.}
}
\description{
This utility function translates a user-supplied string into coordinates for
legend location.
}
