% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{matrixize}
\alias{matrixize}
\title{Standardize Matrix}
\usage{
matrixize(dat)
}
\arguments{
\item{dat}{Omic data matrix or matrix-like object with rows corresponding to
probes and columns to samples.}
}
\description{
This utility function takes data objects from \code{limma}, \code{edgeR}, or
\code{DESeq2} pipelines and outputs a standard probe by sample matrix. Raw counts
are log2-CPM transformed with a warning.
}
