% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_concordance.R
\name{plot_concordance}
\alias{plot_concordance}
\title{Concordance Plot}
\usage{
plot_concordance(dat, label = FALSE, diag = FALSE, method = "MI",
  alpha = NULL, p.adj = NULL, B = 2000, title = NULL,
  legend = "right", hover = FALSE)
}
\arguments{
\item{dat}{A sample by feature data frame or matrix, e.g. of clinical
variables or patient cluster assignments. All columns are converted to
factors with a warning, if possible.}

\item{label}{Print association statistic over tiles?}

\item{diag}{Include principal diagonal of the concordance matrix? Only
advisable if \code{method = "MI"}.}

\item{method}{String specifying which measure of association to
compute. Currently supports \code{"MI"}, \code{"fisher"}, and \code{
"chisq"}. See Details.}

\item{alpha}{Optional significance threshold to impose on association
statistics. Those with \emph{p}-values (optionally adjusted) less than or
equal to \code{alpha} are outlined in black.}

\item{p.adj}{Optional \emph{p}-value adjustment for multiple testing. Options
include \code{"holm"}, \code{"hochberg"}, \code{"hommel"}, \code{
"bonferroni"}, \code{"BH"}, \code{"BY"}, and \code{"fdr"}. See \code{
\link[stats]{p.adjust}}.}

\item{B}{Number of permutations to generate when computing \emph{p}-values
via Monte Carlo simulation. See Details.}

\item{title}{Optional plot title.}

\item{legend}{Legend position. Must be one of \code{"right"}, \code{
"left"}, \code{"top"}, \code{"bottom"}, \code{"topright"}, \code{
"topleft"}, \code{"bottomright"}, or \code{"bottomleft"}.}

\item{hover}{Show correlation coefficient by hovering mouse over the
corresponding tile or circle? If \code{TRUE}, the plot is rendered in HTML
and will either open in your browser's graphic display or appear in the
RStudio viewer.}
}
\description{
This function plots the pairwise concordance between categorical features.
}
\details{
Concordance plots visualize the

When \code{method = "fisher"}, concordance is measured by the negative
logarithm of the test's \emph{p}-value. If this cannot be computed in the
available workspace, then it is estimated via simulation using \code{B}
Monte Carlo permutations.

When \code{method = "chisq"}, concordance is measured by the Pearson
chi-squared statistic. This is easier to compute than Fisher exact tests, but
may be inappropriate if BLAH

When \code{method = "MI"}, concordance is measured by the mutual information
statistic. If \code{alpha} is non-\code{NULL}, then \code{p}-values are
estimated via simulation using \code{B} Monte Carlo permutations.
}
\examples{
df <- data.frame(A = sample(1:2, 20, replace = TRUE),
                 B = sample(1:3, 20, replace = TRUE),
                 C = sample(1:4, 20, replace = TRUE))
plot_concordance(df)

}
