% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_consensus.R
\name{plot_consensus}
\alias{plot_consensus}
\title{Consensus Matrix Heatmap}
\usage{
plot_consensus(cc, k, group = NULL, covar = NULL, hclustfun = "average",
  pal_clust = "d3", pal_group = "npg", pal_covar = "Blues",
  pal_tiles = "RdBu", title = NULL)
}
\arguments{
\item{cc}{A list created by a call to \code{
\link[ConsensusClusterPlus]{ConsensusClusterPlus}}.}

\item{k}{Integer specifying number of clusters to visualize.}

\item{group}{Optional character or factor vector of length equal to sample
size. Alternatively, a data frame or list of such vectors, optionally
named.}

\item{covar}{Optional continuous covariate of length equal to sample size.
Alternatively, a data frame or list of such vectors, optionally named.}

\item{hclustfun}{The agglomeration method to be used for hierarchical
clustering. Supports any method available in \code{\link[stats]{hclust}}.}

\item{pal_clust}{String specifying the color palette to use for cluster
assignments. Options include \code{"ggplot"}, all qualitative color schemes
available in \code{RColorBrewer}, and the complete collection of \code{
\href{http://bit.ly/2bxnuGB}{ggsci}} palettes.}

\item{pal_group}{String specifying the color palette to use if \code{group}
is non-\code{NULL}, or a vector of such strings with length equal to the
number of vectors passed to \code{group}. Options are the same as those for
\code{pal_clust}. Alternatively, a character vector of colors with length
equal to the cumulative number of levels in \code{group}.}

\item{pal_covar}{String specifying the color palette to use if \code{covar}
is non-\code{NULL}, or a vector of such strings with length equal to the
number of vectors passed to \code{covar}. Options include all sequential
color schemes available in \code{RColorBrewer}. Alternatively, a
character vector of colors representing a smooth gradient, or a list of
such vectors with length equal to the number of continuous variables to
visualize.}

\item{pal_tiles}{String specifying the color palette to use for heatmap
tiles. Options include all diverging color schemes available in \code{
RColorBrewer}. Alternatively, a character vector of at least two colors.}

\item{title}{Optional plot title.}
}
\description{
This function displays pairwise consensus cluster distances as a heatmap.
}
\details{
Consensus clustering is a resampling procedure to evaluate cluster stability.
A user-specified proportion of samples are held out on each run of the
algorithm to test how often the remaining samples do or do not cluster
together. The result is a square consensus matrix for each value of cluster
numbers \emph{k}. Each cell of the matrix \code{mat[i, j]} represents the
proportion of all runs including samples \code{i} and \code{j} in which the
two were clustered together.

\code{plot_consensus} converts a consensus matrix into a distance matrix by
taking the complement of all values and visualising the result as a heatmap.
These figures are similar to those presented in the original consensus
cluster paper by Monti et al.
}
\examples{
# Load libraries
library(ConsensusClusterPlus)

# Import, filter, transform data
data(airway)
mat <- assay(airway)
mads <- apply(mat, 1, mad)
keep <- order(mads, decreasing = TRUE)[seq_len(1000)]
mat <- mat[keep, ]
mat <- log2((mat + 1L) / 1e6L)

# Run consensus clustering, plot results
cc <- ConsensusClusterPlus(mat)
plot_consensus(cc, k = 2)

}
\references{
Monti, S., Tamayo, P., Mesirov, J., & Golub, T. (2003).
\href{http://link.springer.com/article/10.1023/A:1023949509487}{Consensus
Clustering: A Resampling-Based Method for Class Discovery and Visualization
of Gene Expression Microarray Data}. \emph{Machine Learning}, \emph{52}:
91-118.
}
