% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_covar.R
\name{plot_covar}
\alias{plot_covar}
\title{Plot associations between omic and clinical data}
\usage{
plot_covar(dat, clin, index = "Sample", block = NULL, n.pc = 10,
  main = NULL, hover = FALSE)
}
\arguments{
\item{dat}{Omic data matrix with rows corresponding to probes and columns
to samples. For best results, data should be normalized and filtered in
preparation for PCA.}

\item{clin}{Data frame with rows correponding to samples and columns to
technical and/or biological covariates to test for associations with omic data.}

\item{index}{String specifying the name of the column in \code{clin} containing
sample names. Only samples in \code{index} with matching column names in
\code{dat} will be considered.}

\item{block}{String specifying the name of the column in which to find the
blocking variable, should one be accounted for. See Details.}

\item{n.pc}{Number of principal components to include in the figure.}

\item{main}{Optional plot title.}

\item{hover}{Show \emph{p}-values by hovering mouse over tiles? If \code{TRUE},
the plot is rendered in HTML and will either open in your browser's graphic
display or appear in the RStudio viewer.}
}
\description{
This function creates a heatmap visualizing the strength of associations
between the principal components of an omic data matrix and a set of
technical and/or biological covariates.
}
\details{
Strength of association is measured using -log10 \emph{p}-values. When
\code{block = NULL}, significance is derived from either Pearson correlation
tests (for continuous features) or one-way ANOVA \emph{F}-tests (for categorical
features).

An optional blocking variable may be provided if samples violate the assumption
of independence, e.g. for studies in which subjects are observed at multiple
time points. If a blocking variable is identified, it will be factored into all
subsequent measures of association. Significance is then evaluated using Pearson
partial correlation tests (for continuous features) or repeated measures ANOVA
\emph{F}-tests (for categorical features). When supplying a blocking variable,
be sure to check that it is not confounded with other features. For instance,
biological covariates like sex and age are usually nested within subject, while
subject may be nested within other features like batch or treatment group.
}
\examples{
data(Nevins)
mat <- exprs(Nevins)
clin <- pData(Nevins)
clin$Sample <- rownames(clin)
plot_covar(mat, clin)

}
