% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Omic Heatmap}
\usage{
plot_heatmap(dat, feat = NULL, col = "rb", main = NULL)
}
\arguments{
\item{dat}{Omic data matrix or matrix-like object with rows corresponding to
probes and columns to samples.}

\item{feat}{Optional character, factor, numeric, or logical vector of length
equal to sample size. Alternatively, a data frame or list of such vectors,
optionally named. Values are used to color one or several annotation tracks
atop the heatmap.}

\item{col}{Color scheme to use for heatmap tiles. Options are \code{"rb"} (for
red to blue gradient) or \code{"gr"} (for green to red gradient).}

\item{main}{Optional plot title.}
}
\description{
This function visualizes a probe by sample matrix as a heatmap.
}
\details{
Heatmaps are a common and intuitive way to display the values of an omic data
matrix, especially after top probes have been selected through statistical testing.
Hierarchical clustering dendrograms cluster both the rows and the columns,
revealing latent structure in the data. Annotation tracks atop the figure may be
used to investigate associations with phenotypic features.
}
\examples{
mat <- matrix(rnorm(1000 * 10), nrow = 1000, ncol = 10)
grp <- rep(c("A", "B"), each = 5)
plot_heatmap(mat, feat = grp)

}
