% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ica.R
\name{plot_ica}
\alias{plot_ica}
\title{ICA Plot}
\usage{
plot_ica(dat, group = NULL, covar = NULL, top = NULL, dims = c(1L,
  2L), label = FALSE, pal_group = "npg", pal_covar = "Blues",
  size = NULL, alpha = NULL, title = "ICA", legend = "right",
  hover = FALSE, D3 = FALSE)
}
\arguments{
\item{dat}{Omic data matrix or matrix-like object with rows corresponding to
probes and columns to samples. It is strongly recommended that data be
filtered and normalized prior to plotting. Raw counts stored in \code{
\link[edgeR]{DGEList}} or \code{\link[DESeq2]{DESeqDataSet}} objects are
automatically extracted and transformed to the log2-CPM scale, with a
warning.}

\item{group}{Optional character or factor vector of length equal to sample
size, or up to two such vectors organized into a list or data frame. Supply
legend title(s) by passing a named list or data frame.}

\item{covar}{Optional continuous covariate. If non-\code{NULL}, then plot can
render at most one \code{group} variable. Supply legend title by passing
a named list or data frame.}

\item{top}{Optional number (if > 1) or proportion (if < 1) of most variable
probes to be used for ICA.}

\item{dims}{Vector specifying which independent components to plot. Must be 
of length two unless \code{D3 = TRUE}.}

\item{label}{Label data points by sample name? Defaults to \code{FALSE}
unless \code{group} and \code{covar} are both \code{NULL}. If \code{TRUE},
then plot can render at most one phenotypic feature.}

\item{pal_group}{String specifying the color palette to use if \code{group}
is non-\code{NULL}, or a vector of such strings with length equal to the
number of vectors passed to \code{group}. Options include \code{"ggplot"}, 
all qualitative color schemes available in \code{\href{
https://bit.ly/2ipuEjn}{RColorBrewer}}, and the complete collection of 
\code{\href{http://bit.ly/2bxnuGB}{ggsci}} palettes. Alternatively, a 
character vector of colors with length equal to the cumulative number of 
levels in \code{group}.}

\item{pal_covar}{String specifying the color palette to use if \code{covar}
is non-\code{NULL}, or a vector of such strings with length equal to the
number of vectors passed to \code{covar}. Options include the complete 
collection of \code{\href{https://bit.ly/2n7D6tF}{viridis}} palettes, as 
well as all sequential color schemes available in \code{\href{
https://bit.ly/2ipuEjn}{RColorBrewer}}. Alternatively, a character vector 
of colors representing a smooth gradient, or a list of such vectors with 
length equal to the number of continuous variables to visualize.}

\item{size}{Point size.}

\item{alpha}{Point transparency.}

\item{title}{Optional plot title.}

\item{legend}{Legend position. Must be one of \code{"bottom"}, \code{"left"},
\code{"top"}, \code{"right"}, \code{"bottomright"}, \code{"bottomleft"},
\code{"topleft"}, or \code{"topright"}.}

\item{hover}{Show sample name by hovering mouse over data point? If \code{
TRUE}, the plot is rendered in HTML and will either open in your browser's
graphic display or appear in the RStudio viewer.}

\item{D3}{Render plot in three dimensions?}
}
\description{
This function plots a low-dimensional projection of an omic data matrix using
independent component analysis.
}
\details{
This function plots the samples of an omic data matrix in a two- or
three-dimensional independent component subspace. ICA is an easy and
popular projection method that can aid in identifying clusters, detecting
outliers, and visualizing the latent structure of a dataset.

ICA is a general method for separating a multivariate signal into additive
subcomponents. ICA algorithms differ in their objective functions and 
optimization methods. \code{plot_ica} relies on Cardoso's JADE ICA algorithm,
as implemented in the \code{\link[JADE]{JADE}} package. 

By default, \code{plot_ica} decomposes the complete \code{dat} matrix. 
Limit the ICA to only the most variable probes by using the \code{top} 
argument.
}
\examples{
n <- 10L
p <- 1000L
x <- matrix(rnorm(n * p), ncol = n)
plot_ica(x)

}
\references{
Miettinen, J., Nordhausen, K. & Taskinen, S. (2017). 
\href{https://www.jstatsoft.org/article/view/v076i02}{Blind Source Separation 
Based on Joint Diagonalization in R: The Packages JADE and BSSasymp}. \emph{
Journal of Statistical Software}, \emph{76}: 1–31.
}
\seealso{
\code{\link[JADE]{JADE}}, \code{\link{plot_pca}}, \code{\link{plot_tsne}}
}
