% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lle.R
\name{plot_lle}
\alias{plot_lle}
\title{LLE Plot}
\usage{
plot_lle(dat, group = NULL, covar = NULL, k = ncol(dat)/4L,
  top = NULL, dims = c(1L, 2L), label = FALSE, pal_group = "npg",
  pal_covar = "Blues", title = NULL, legend = "right",
  hover = FALSE, D3 = FALSE)
}
\arguments{
\item{dat}{Omic data matrix or matrix-like object with rows corresponding to
probes and columns to samples. It is strongly recommended that data be
filtered and normalized prior to plotting. Raw counts stored in \code{
\link[edgeR]{DGEList}} or \code{\link[DESeq2]{DESeqDataSet}} objects are
automatically extracted and transformed to the log2-CPM scale, with a
warning.}

\item{group}{Optional character or factor vector of length equal to sample
size, or up to two such vectors organized into a list or data frame. Supply
legend title(s) by passing a named list or data frame.}

\item{covar}{Optional continuous covariate. If non-\code{NULL}, then plot can
render at most one \code{group} variable. Supply legend title by passing
a named list or data frame.}

\item{k}{How many nearest neighbors should the algorithm consider when
building projections? See Details.}

\item{top}{Optional number (if > 1) or proportion (if < 1) of most variable
probes to be used for LLE.}

\item{dims}{Vector specifying which principal components to plot. Must be of
length two unless \code{D3 = TRUE}.}

\item{label}{Label data points by sample name? Defaults to \code{FALSE}
unless \code{group} and \code{covar} are both \code{NULL}. If \code{TRUE},
then plot can render at most one phenotypic feature.}

\item{pal_group}{String specifying the color palette to use if \code{group}
is non-\code{NULL}, or a vector of such strings with length equal to the
number of vectors passed to \code{group}. Options include \code{"ggplot"},
all qualitative color schemes available in \code{RColorBrewer}, and the
complete collection of \code{\href{http://bit.ly/2bxnuGB}{ggsci}} palettes.
Alternatively, a character vector of colors with length equal to the
cumulative number of levels in \code{group}.}

\item{pal_covar}{String specifying the color palette to use if \code{covar}
is non-\code{NULL}, or a vector of such strings with length equal to the
number of vectors passed to \code{covar}. Options include all sequential
color schemes available in \code{RColorBrewer}. Alternatively, a
character vector of colors representing a smooth gradient, or a list of
such vectors with length equal to the number of continuous variables to
visualize.}

\item{title}{Optional plot title.}

\item{legend}{Legend position. Must be one of \code{"bottom"}, \code{"left"},
\code{"top"}, \code{"right"}, \code{"bottomright"}, \code{"bottomleft"},
\code{"topleft"}, or \code{"topright"}.}

\item{hover}{Show sample name by hovering mouse over data point? If \code{
TRUE}, the plot is rendered in HTML and will either open in your browser's
graphic display or appear in the RStudio viewer.}

\item{D3}{Render plot in three dimensions?}
}
\description{
This function plots a low-dimensional projection of an omic data matrix using
locally linear embedding.
}
\details{
Locally linear embedding is a feature extraction method that projects a
high-dimensional dataset in two or three dimensions. It prioritizes local
over global structure, and is designed to capture nonlinearities that would
be impossible to detect using PCA or classical MDS.

By default, the number of nearest neighbors \emph{k} used in the LLE
projection is fixed at one quarter the sample size. This is an arbitrary
value that may be unsuitable for some datasets. The parameter may be
optimized using Kayo's method (2006). See See \code{lle::\link[lle]{calc_k}}.
}
\examples{
mat <- matrix(rnorm(1000 * 5), nrow = 1000, ncol = 5)
plot_lle(mat)

library(DESeq2)
dds <- makeExampleDESeqDataSet()
dds <- rlog(dds)
plot_lle(dds, group = colData(dds)$condition)

}
\references{
Roweis, S.T. & Saul, L.K. (2000).
\href{http://science.sciencemag.org/content/290/5500/2323}{Nonlinear
Dimensionality Reduction by Locally Linear Embedding}. \emph{Science,
290}(5500): 2323-2326.

Kayo, O. (2006).
\href{http://jultika.oulu.fi/files/isbn9514280415.pdf}{Locally Linear
Embedding Algorithm: Extensions and Applications} (Unpublished doctoral
dissertation). University of Oulu, Finland.
}
\seealso{
\code{\link{plot_pca}}, \code{\link{plot_tsne}}
}
