% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_md.R
\name{plot_md}
\alias{plot_md}
\title{Mean-Difference Plot}
\usage{
plot_md(dat, fdr = 0.05, ptsize = 0.25, main = NULL, legend = "outside",
  hover = FALSE, probes = NULL)
}
\arguments{
\item{dat}{Data frame or matrix representing the results of a test for
differential expression or methylation, such as the output of a call to
\code{limma::topTable, edgeR::topTags}, or \code{DESeq2::results}.
Alternatively, any object with columns for log fold changes and FDR.}

\item{fdr}{Threshold for declaring a probe differentially expressed or methylated.}

\item{ptsize}{Size of data points in the plot.}

\item{main}{Optional plot title.}

\item{legend}{Legend position. Must be one of \code{"outside",
"bottomleft", "bottomright", "topleft",} or \code{"topright"}.}

\item{hover}{Show probe name by hovering mouse over data point? If \code{TRUE},
the plot is rendered in HTML and will either open in your browser's graphic
display or appear in the RStudio viewer.}

\item{probes}{String specifying the name of the column in which to find the probe
identifiers, assuming they aren't \code{rownames(dat)}. Only relevant if
\code{hover = TRUE}.}
}
\description{
This function plots probewise means against the log2 fold changes of a
differential expression or methylation test.
}
\details{
MD plots (also known as "Bland-Altman plots" or "MA plots") visualize the
relationship between a probe's mean value and its log2 fold change for a given
test of differential expression/methylation. Probes are colored to distinguish
between those that do and do not meet a user-defined FDR threshold. These
figures help to evaluate the symmetry, magnitude, and significance of effects
for a given experiment.
}
\examples{
df <- data.frame(logFC   = c(rnorm(50, 0, 10), rnorm(4950)),
                 AvgExpr = rowMeans(matrix(rnorm(5000), nrow = 5000, ncol = 10)),
                 p.value = pnorm(-abs(logFC)),
                 FDR     = p.adjust(p.value, method = "fdr"))
plot_md(df)

library(limma)
DE_genes <- cbind(matrix(rnorm(250, 5, 1), nrow = 50, ncol = 5),
                  matrix(rnorm(250), nrow = 50, ncol = 5))
mat <- rbind(DE_genes, matrix(rnorm(45500), nrow = 4550, ncol = 10))
treat <- gl(n = 2, k = 5, labels = c("A", "B"))
des <- model.matrix(~ treat)
fit <- eBayes(lmFit(mat, des))
top <- topTable(fit, number = Inf)
plot_md(top)

}
