% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mv.R
\name{plot_mv}
\alias{plot_mv}
\title{Mean-Variance Plot}
\usage{
plot_mv(dat, trans = NULL, zero.weights = FALSE, ptsize = 0.25,
  main = NULL, legend = "outside", hover = FALSE)
}
\arguments{
\item{dat}{Either an omic data matrix with rows corresponding to probes and
columns to samples, or an object of class \code{\link[limma]{MArrayLM}} as
created by a call to \code{\link[limma]{lmFit}} or \code{\link[limma]{eBayes}}.
If \code{dat} is a matrix, then data are presumed to be normalized prior to
visualization. Any matrix-like object that can be processed by
\code{\link[limma]{getEAWP}} is also acceptable.}

\item{trans}{Data transformation to be applied to probewise standard deviations
if \code{dat} is a matrix. Must be one of either \code{"log"} or \code{"sqrt"}.
See Details.}

\item{zero.weights}{Should spots with zero or negative weights be plotted?}

\item{ptsize}{Size of data points in the plot.}

\item{main}{Optional plot title.}

\item{legend}{Legend position. Must be one of \code{"outside", "bottomleft",
"bottomright", "topleft",} or \code{"topright"}. Only relevant if \code{dat}
is an \code{MArrayLM} object created by a call to \code{eBayes} with
\code{trend = TRUE}.}

\item{hover}{Show probe name by hovering mouse over data point? If \code{TRUE},
the plot is rendered in HTML and will either open in your browser's graphic
display or appear in the RStudio viewer. Probe names are extracted from
\code{dat}.}
}
\description{
This function visualizes the mean-variance relationship of omic data before
or after linear modeling.
}
\details{
Mean-variance plots are a quick and easy way to visualize the relationship
between the first two moments of an omic data distribution. When used prior to
modeling, they may help better understand the internal structure of the data
and inspect for potential outliers. When used after modeling, they can be
useful in evaluating the assumptions of the regression. \code{plot_mv} fits
a smooth curve to the points using a generalized additive model.

If \code{dat} is a matrix, then the appropriate data transformation for probewise
standard deviations must be specified. \code{trans = "log"} is recommended for
microarrays or any other platform in which data are approximately log-normally
distributed. \code{trans = "sqrt"} is recommended for sequencing and count data.

If \code{dat} is an \code{MArrayLM} object, then the y-axis will be log2
transformed probewise residual standard deviations. If standard errors were
moderated with \code{eBayes}, then prior variance will also be plotted as either
a horizontal line or a smooth curve, depending on whether a global or
intensity-dependent prior was used. See \code{\link[limma]{squeezeVar}} for more
details.
}
\examples{
mat <- matrix(rnorm(1000 * 10), nrow = 1000, ncol = 10)
plot_mv(mat, trans = "log")

library(limma)
grp <- rep(c("ctl", "trt"), each = 5)
des <- model.matrix(~ grp)
fit <- eBayes(lmFit(mat, des))
plot_mv(fit)

}
