% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca.R
\name{plot_PCA}
\alias{plot_PCA}
\title{Plot the first two principal components of an omic data matrix}
\usage{
plot_PCA(dat, group = NULL, label = FALSE, main = NULL,
  legend = "outside")
}
\arguments{
\item{dat}{Omic data matrix with rows corresponding to probes and columns 
to samples.}

\item{group}{Optional factor or character vector of length equal to sample size. 
Levels are used to color and shape points.}

\item{label}{Label data points by sample name? Defaults to FALSE unless 
\code{group = NULL}.}

\item{main}{Optional plot title.}

\item{legend}{Legend position. Must be one of \code{"outside", 
"bottomleft", "bottomright", "topleft",} or \code{"topright"}.}
}
\description{
Plot the first two principal components of an omic data matrix
}
\details{
This function plots the first two principal components of an omic data 
matrix. The numbers printed in the x- and y-axis labels indicate the 
percentage of variance explained by each component. PCA is an easy and 
popular method for unsupervised cluster detection. It can also aid in 
spotting potential outliers and generally  help to visualize the latent 
structure of a data set.
}
\examples{
mat <- matrix(rnorm(5000), nrow = 1000, ncol = 5)
plot_pca(mat)

library(DESeq2)
mat <- cbind(matrix(rnbinom(5000, mu = 4, size = 1), nrow = 1000, ncol = 5),
             matrix(rnbinom(5000, mu = 4, size = 10), nrow = 1000, ncol = 5))
mat <- rlog(mat)
batch <- rep(c("A", "B"), each = 5)
plot_pca(mat, group = batch)

}

