% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_qq.R
\name{plot_qq}
\alias{plot_qq}
\title{Create QQ plot of expected vs. observed \emph{p}-values}
\usage{
plot_qq(dat, ptsize = 0.25, main = NULL, legend = "outside",
  hover = FALSE, probes = NULL)
}
\arguments{
\item{dat}{Data frame or matrix representing the results of a test for
differential expression or methylation, such as the output of a call to
\code{limma::topTable, edgeR::topTags}, or \code{DESeq2::results}.
Alternatively, any object with a column for \emph{p}-values.}

\item{ptsize}{Size of data points in the plot.}

\item{main}{Optional plot title.}

\item{legend}{Legend position. Must be one of \code{"outside",
"bottomleft", "bottomright", "topleft",} or \code{"topright"}.}

\item{hover}{Show probe name by hovering mouse over data point? If \code{TRUE},
the plot is rendered in HTML and will either open in your browser's graphic
display or appear in the RStudio viewer.}

\item{probes}{String specifying the colname in which to find the probe names.
Only relevant if \code{hover = TRUE}.}
}
\description{
Create QQ plot of expected vs. observed \emph{p}-values
}
\details{
This function displays a scatterplot of the expected vs. observed quantiles
of a \emph{p}-value distribution following a -log10 transform. If the black points
deviate too sharply from the red line, especially at low expected values of
-log10(\emph{p}), then it suggests that one or several statistical assumptions
upon which the test was based have probably been violated.
}
\examples{
df <- data.frame(p.value = runif(10000))
plot_qq(df)

}

