% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_quantiles.R
\name{plot_quantiles}
\alias{plot_quantiles}
\title{Plot Quantiles}
\usage{
plot_quantiles(x, y, method = "QQ", pts = 1000, main = NULL,
  xlab = NULL, ylab = NULL)
}
\arguments{
\item{x}{Vector of numeric values.}

\item{y}{Second vector of numeric values.}

\item{method}{Plot quantiles against quantiles (\code{method = "QQ"}) or mean
quantiles against difference in quantiles (\code{method = "MD"})?}

\item{pts}{Number of points to plot.}

\item{main}{Optional plot title.}

\item{xlab}{Optional x-axis label.}

\item{ylab}{Optional y-axis label.}
}
\description{
This function plots the quantiles for two vectors against each other as either a
QQ or an MD plot.
}
\details{
QQ and MD plots visualize the relationship between two numeric vectors. They are
a quick and easy alternative to scatterplots when distributions are of unequal
length.
}
\examples{
x1 <- rnorm(100)
x2 <- runif(500)
plot_quantiles(x1, x2)
plot_quantiles(x1, x2, method = "MD")

}
