% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_roc.R
\name{plot_roc}
\alias{plot_roc}
\title{Reicever Operating Characteristic Curve(s)}
\usage{
plot_roc(obs, pred, pal_curves = "npg", title = NULL, leg.txt = NULL,
  legend = "bottomright", hover = FALSE)
}
\arguments{
\item{obs}{Vector of observed outcomes. Must be dichotomous. Can be logical,
numeric, character, or factor. If numeric, \code{obs} must be coded \code{
1} or \code{0}. If character or factor, a warning will be issued clarifying
that the first level is assumed to be the reference.}

\item{pred}{Vector of predicted values, or several such vectors organized
into a data frame or list, optionally named. Must be numeric. Common
examples include the probabilities output by a logistic model, or the
expression levels of a particular biomarker.}

\item{pal_curves}{String specifying the color palette to use when plotting
multiple curves. Options include \code{"ggplot"}, as well as the complete
collection of \code{
\href{https://cran.r-project.org/web/packages/ggsci/vignettes/ggsci.html}{
ggsci}} palettes, which can be identified by name (e.g., \code{"npg"},
\code{"aaas"}, etc.). Alternatively, a character vector of colors with
length equal to the number of vectors in \code{pred}.}

\item{title}{Optional plot title.}

\item{leg.txt}{Optional legend title.}

\item{legend}{Legend position. Must be one of \code{"right"}, \code{
"left"}, \code{"top"}, \code{"bottom"}, \code{"bottomright"},
\code{"bottomleft"}, \code{"topright"}, or \code{"topleft"}.}

\item{hover}{Show predictor name by hovering mouse over ROC curve? If \code{
TRUE}, the plot is rendered in HTML and will either open in your browser's
graphic display or appear in the RStudio viewer.}
}
\description{
This functions plots ROC curves for one or several classifiers.
}
\details{
ROC curves plot the false positive rate (i.e., 1 - specificity) against the
true positive rate (i.e., sensitivity) for a given classifier and vector of
observations. The area under the ROC curve (AUC) is a common performance
metric for binary classifiers. The grey diagonal line across the plot
represents the performance of a theoretical random classifier.
}
\examples{
y <- rbinom(1000, size = 1, prob = 0.5)
x1 <- rnorm(1000, mean = y)
plot_roc(obs = y, pred = x1)

x2 <- rnorm(1000, mean = y, sd = 2)
plot_roc(obs = y, pred = list("x1" = x1, "x2" = x2))

}
