% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sm.R
\name{plot_sm}
\alias{plot_sm}
\title{Similarity Matrix Heatmap}
\usage{
plot_sm(dat, feat = NULL, main = NULL)
}
\arguments{
\item{dat}{Omic data matrix or matrix-like object with rows corresponding to
probes and columns to samples.}

\item{feat}{Optional character, factor, numeric, or logical vector of length
equal to sample size. Alternatively, a data frame or list of such vectors,
optionally named. Values are used to color one or several annotation tracks
atop the heatmap.}

\item{main}{Optional plot title.}
}
\description{
This function displays pairwise Euclidean distances between samples as a
heatmap.
}
\details{
Similarity matrices are a valuable tool for exploratory data analysis. A
hierarchical clustering dendrogram atop the figure helps identify potential
outliers and/or clusters in the data. Annotation tracks can help investigate
associations with phenotypic features.
}
\examples{
mat <- matrix(rnorm(5000), nrow = 1000, ncol = 5)
plot_sm(mat, main = "Nothin' Doin'")

library(DESeq2)
mat <- cbind(matrix(rnbinom(5000, mu = 4, size = 1), nrow = 1000, ncol = 5),
             matrix(rnbinom(5000, mu = 4, size = 10), nrow = 1000, ncol = 5))
mat <- rlog(mat)
grp <- rep(c("A", "B"), each = 5)
plot_sm(mat, feat = grp, main = "Somethin' Cookin'")

}
