% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_som.R
\name{plot_som}
\alias{plot_som}
\title{SOM}
\usage{
plot_som(dat, type = "model", design = NULL, coef = NULL, sample = NULL,
  top = 0.5, grid_dim = NULL, topo = "hexagonal", neighb = "bubble",
  rlen = 1000, parallel = TRUE, pal_tiles = NULL, title = NULL,
  legend = "right", hover = FALSE, export = TRUE, ...)
}
\arguments{
\item{dat}{Either a probe by sample omic data matrix or an object of class
\code{kohonen}. See Details.}

\item{type}{What should the plot visualize? Options include \code{"model"},
in which case \code{design} and \code{coef} must be supplied; \code{
"sample"}, in which case the \code{sample} must be specified; \code{
"distance"}; \code{"counts"}; or \code{"train"}. See Details.}

\item{design}{Design matrix for linear model with rows corresponding to
samples and columns to model coefficients. Only relevant if \code{type =
"model"}.}

\item{coef}{Column number or name specifying which coefficient or contrast
of the linear model is of interest. Only relevant if \code{type = "model"}.}

\item{sample}{Column number or name specifying which sample should be
plotted. Only relevant if \code{type = "sample"}.}

\item{top}{Optional number (if > 1) or proportion (if < 1) of most variable
probes to be used for SOM.}

\item{grid_dim}{Integer length for x- and y-axis of the SOM grid. Total node
count is given by \code{grid_dim ^ 2}. If \code{NULL}, dimensionality will
be calculated so as to put approximately 10 probes in each node, presuming
they are uniformly distributed across the map.}

\item{topo}{Train SOM using \code{"rectangular"} or \code{"hexagonal"}
topology?}

\item{neighb}{Train SOM using a \code{"bubble"} or \code{"gaussian"}
neighborhood function?}

\item{rlen}{Run length for SOM training.}

\item{parallel}{Allow for parallel computation of SOM?}

\item{pal_tiles}{String specifying the color palette to use for heatmap
nodes. Defaults to \code{"Spectral"} if \code{type = "model"} or \code{
"sample"}; \code{"Greys"} if \code{type = "distance"}; and \code{"Blues"}
if \code{type = "counts"}. Any divergent or sequential palette in \code{
RColorBrewer} is acceptable. Alternatively, any vector of recognized colors
may be supplied.}

\item{title}{Optional plot title.}

\item{legend}{Legend position. Must be one of \code{"bottom"}, \code{"left"},
\code{"top"}, \code{"right"}, \code{"bottomright"}, \code{"bottomleft"},
\code{"topleft"}, or \code{"topright"}.}

\item{hover}{Show constituent probes by hovering mouse over module tile? If
\code{TRUE}, the plot is rendered in HTML and will either open in your
browser's graphic display or appear in the RStudio viewer.}

\item{export}{Return fitted SOM model?}

\item{...}{Extra arguments to be passed to \code{\link[limma]{lmFit}} if
\code{type = "model"}.}
}
\description{
This function plots the output of a self-organizing map.
}
\details{
A self-organizing map (SOM) is a type of neural network used for
dimensionality reduction and data visualization. SOMs come in many flavors,
but the kind used here clusters probes together in an unsupervised manner to
identify functionally relevant gene sets, examine patterns across samples,
and explore hypotheses. See the references below for more details on SOM
algorithms and their applications for omic research.

If \code{type = "model"}, then a linear model is fit to the codebook vectors.
Units are colored by each node's log fold change associated with \code{coef}.

If \code{type = "sample"}, then the plot depicts that sample's expression
profile across the SOM.

If \code{type = "distance"}, then the figure renders the SOM's U-matrix,
which represents the Euclidean distance between codebook vectors for
neighboring units. This can be used to inspect for clusters and borders
in the SOM space.

If \code{type = "counts"}, then units are colored by each node's probe
count. This distribution should ideally be nearly uniform. A large number of
empty units suggests that \code{grid_dim} should be reduced; an uneven spread
across the units suggests that \code{grid_dim} should be increased.

If \code{type = "train"}, then the plot displays the SOM's learning curve
over its \code{rlen} training iterations.

Depending on the data's size, the model's parameters, and your computational
resources, it may take a considerable amount of time to fit a SOM to an
omic matrix. That is why \code{plot_som} returns the SOM object by default,
which may in turn be reused in subsequent calls to the function to explore
alternative aspects of the map. SOMs are trained using the batch algorithm
of \code{kohonen::\link[kohonen]{som}}. This may be executed in parallel for
faster mapping. See that package's documentation for more details.
}
\examples{
mat <- matrix(rnorm(1000 * 5), nrow = 1000, ncol = 5)
plot_som(mat, type = "sample", sample = 1)

}
\references{
Kohonen, T. (1995). \emph{Self-Organizing Maps}. Berlin: Springer-Verlag.

Nikkilä, J., Törönen, P., Kaski, S., Venna, J., Castrén, E., & Wong, G. (2002).
\href{http://www.sciencedirect.com/science/article/pii/S0893608002000709}{
Analysis and visualization of gene expression data using Self-Organizing
Maps}. \emph{Neural Networks}, \emph{15}(8-9): 953-966.

Wehrens, R. & Buydens, L.M.C. (2007). \href{http://bit.ly/2tBFE6R}{Self- and
Super-organizing Maps in R: The kohonen Package} \emph{J. Stat. Softw.},
\emph{21}(5).

Wirth, H. (2012). \href{http://bit.ly/2tGq6iY}{Analysis of large-scale
molecular biological data using self-organizing maps}. Dissertation thesis,
University of Leipzig.
}
