% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_taco.R
\name{plot_taco}
\alias{plot_taco}
\title{Taco Plot}
\usage{
plot_taco(dat, fdr = 0.05, main = NULL, legend = "outside",
  probes = NULL)
}
\arguments{
\item{dat}{Data frame or matrix representing the results of a test for
differential expression or methylation, such as the output of a call to
\code{limma::topTable, edgeR::topTags}, or \code{DESeq2::results}.
Alternatively, any object with columns for \emph{p}-values, log fold changes,
and FDR.}

\item{fdr}{Threshold for declaring a probe differentially expressed/methylated.}

\item{main}{Optional plot title.}

\item{legend}{Legend position. Must be one of \code{"outside",
"bottomleft", "bottomright", "topleft",} or \code{"topright"}.}

\item{probes}{String specifying the name of the column in which to find the probe
identifiers, assuming they aren't \code{rownames(dat)}.}
}
\description{
This function plots log2 fold changes against -log10 \emph{p}-values against
probewise means for a given test of differential expression/methylation.
}
\details{
A taco plot combines the elements of a volcano plot and an MD plot into a single
three-dimensional figure. Probes are colored to distinguish between those that
do and do not meet a user-defined FDR threshold. These figures help to evaluate
the symmetry, magnitude, and significance of effects for a given experiment.
}
\examples{
df <- data.frame(logFC   = c(rnorm(50, 0, 10), rnorm(4950)),
                 AvgExpr = rowMeans(matrix(rnorm(5000), nrow = 5000, ncol = 10)),
                 p.value = pnorm(-abs(logFC)),
                 FDR     = p.adjust(p.value, method = 'fdr'))
plot_taco(df)

library(limma)
DE_genes <- cbind(matrix(rnorm(250, 5, 1), nrow = 50, ncol = 5),
                  matrix(rnorm(250), nrow = 50, ncol = 5))
mat <- rbind(DE_genes, matrix(rnorm(45500), nrow = 4550, ncol = 10))
treat <- gl(n = 2, k = 5, labels = c("A", "B"))
des <- model.matrix(~ treat)
fit <- eBayes(lmFit(mat, des))
top <- topTable(fit, number = Inf)
plot_taco(top)

}
