% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{probe_alpha}
\alias{probe_alpha}
\title{Set Transparency for Probewise Data Points}
\usage{
probe_alpha(df)
}
\arguments{
\item{df}{Data frame to be passed to \code{ggplot}.}
}
\description{
This utility function assigns a reasonable alpha parameter based on data
dimensionality.
}
