% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{probe_ptsize}
\alias{probe_ptsize}
\title{Size Probewise Data Points}
\usage{
probe_ptsize(df)
}
\arguments{
\item{df}{Data frame to be passed to \code{ggplot}.}
}
\description{
This utility function assigns a reasonable point size parameter based on data
dimensionality.
}
