% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pt_alpha}
\alias{pt_alpha}
\title{Set Data Point Transparency}
\usage{
pt_alpha(df)
}
\arguments{
\item{df}{Data frame to be passed to \code{ggplot}.}
}
\description{
This utility function assigns a reasonable alpha parameter based on data
dimensionality.
}
