% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pt_size}
\alias{pt_size}
\title{Size Data Points}
\usage{
pt_size(df)
}
\arguments{
\item{df}{Data frame to be passed to \code{ggplot}.}
}
\description{
This utility function assigns a reasonable point size parameter based on data
dimensionality.
}
