% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sample_alpha}
\alias{sample_alpha}
\title{Set Transparency for Samplewise Data Points}
\usage{
sample_alpha(df)
}
\arguments{
\item{df}{Data frame to be passed to \code{ggplot}.}
}
\description{
This utility function assigns a reasonable alpha parameter based on data
dimensionality.
}
