% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sample_ptsize}
\alias{sample_ptsize}
\title{Size Samplewise Data Points}
\usage{
sample_ptsize(df)
}
\arguments{
\item{df}{Data frame to be passed to \code{ggplot}.}
}
\description{
This utility function assigns a reasonable point size parameter based on data
dimensionality.
}
