% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{track_cols}
\alias{track_cols}
\title{Format Annotation Track Colors}
\usage{
track_cols(features, pal, var_type)
}
\arguments{
\item{features}{List output by \code{\link{format_features}}.}

\item{pal}{Palette name or constituents.}

\item{var_type}{Character string specifying whether features are categorical
or continuous.}
}
\description{
This utility function formats the colors for plotting annotation tracks atop
heatmaps.
}
