% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{var_filt}
\alias{var_filt}
\title{Filter by Variance}
\usage{
var_filt(dat, top, robust)
}
\arguments{
\item{dat}{Omic data matrix.}

\item{top}{Optional number (if > 1) or proportion (if < 1) of top probes to
be used for distance calculations.}

\item{robust}{Use robust variance statistic?}
}
\description{
This utility function filters a matrix by probewise variance or median
absolute deviation.
}
