% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{colorize}
\alias{colorize}
\title{Assign Colors}
\usage{
colorize(pal, var_type, n)
}
\arguments{
\item{pal}{Palette name or constituents.}

\item{var_type}{Character string specifying whether features are \code{
"Categorical"} or \code{"Continuous"}.}

\item{n}{Number of unique groups for which colors must be assigned. Only
relevant if \code{var_type = "Categorical"}.}
}
\description{
This utility function applies a user selected color palette to a ggplot
image.
}
