% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dist_mat}
\alias{dist_mat}
\title{Create Distance Matrix}
\usage{
dist_mat(dat, top, dist)
}
\arguments{
\item{dat}{Omic data matrix or matrix-like object with rows corresponding to
probes and columns to samples.}

\item{top}{Optional number (if > 1) or proportion (if < 1) of top probes to be used
for distance calculations.}

\item{dist}{Distance measure to be used. Currently supports \code{"euclidean",
"pearson", "MI",} or \code{"KLD"}.}
}
\description{
This utility function calculates distance based on a user defined measure
and optionally filters probes by leading log fold change.
}
