% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gg_out}
\alias{gg_out}
\title{Output Image}
\usage{
gg_out(p, hover, loc = NULL)
}
\arguments{
\item{p}{A \code{ggplot2} object.}

\item{hover}{Add text tooltip using plotly?}

\item{loc}{String specifying legend location.}
}
\description{
This utility function locates the figure legend and prints a ggplot or
ggplotly figure.
}
