% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_box.R
\name{plot_box}
\alias{plot_box}
\title{Box Plots by Sample}
\usage{
plot_box(dat, group = NULL, type = NULL, ylab = NULL, main = NULL,
  legend = "outside", hover = FALSE)
}
\arguments{
\item{dat}{Omic data matrix with rows corresponding to probes and columns
to samples.}

\item{group}{Optional character or factor vector of length equal to sample size.
Levels are used to color box plots. If supplied, legend title defaults to
"Group". Override this feature by passing a named list instead.}

\item{ylab}{Optional label for y-axis.}

\item{main}{Optional plot title.}

\item{legend}{Legend position. Must be one of \code{"outside",
"bottomleft", "bottomright", "topleft",} or \code{"topright"}.}

\item{hover}{Show sample name by hovering mouse over data point? If \code{TRUE},
the plot is rendered in HTML and will either open in your browser's graphic
display or appear in the RStudio viewer.}
}
\description{
This function displays each sample's omic data distribution as a box and
whicker plot.
}
\details{
Box plots are an intuitive way to visualize an omic data distribution. They are
especially helpful when contrasting pre- and post-normalization matrices.
\code{plot_box} may additionally be used to inspect for batch effects
or associations with phenotypic factors by using the \code{group} argument.
}
\examples{
mat <- matrix(rnorm(1000 * 5), nrow = 1000, ncol = 5)
plot_box(mat, ylab = "Normalized Expression")

library(edgeR)
mat <- cbind(matrix(rnbinom(1000 * 5, mu = 4, size = 1), nrow = 1000, ncol = 5),
             matrix(rnbinom(1000 * 5, mu = 4, size = 5), nrow = 1000, ncol = 5))
mat <- calcNormFactors(DGEList(mat))
mat <- cpm(mat, log = TRUE)
batch <- rep(c("A", "B"), each = 5)
plot_box(mat, group = batch, ylab = "Normalized Counts")

}
