% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_concordance.R
\name{plot_concordance}
\alias{plot_concordance}
\title{Concordance Plot}
\usage{
plot_concordance(dat, method = "fisher", alpha = NULL, p.adj = NULL,
  sim.p = FALSE, B = 2000L, label = FALSE, diag = FALSE,
  title = "Concordance Plot", legend = "right", hover = FALSE,
  export = FALSE)
}
\arguments{
\item{dat}{A sample by feature data frame or matrix, e.g. of clinical
variables or patient cluster assignments. All columns are converted to
factors with a warning, if possible.}

\item{method}{String specifying which measure of association to
compute. Currently supports \code{"fisher"}, \code{"chisq"}, and \code{
"MI"}. See Details.}

\item{alpha}{Optional significance threshold to impose on association
statistics. Those with \emph{p}-values (optionally adjusted) less than or
equal to \code{alpha} are outlined in black.}

\item{p.adj}{Optional \emph{p}-value adjustment for multiple testing. Options
include \code{"holm"}, \code{"hochberg"}, \code{"hommel"}, \code{
"bonferroni"}, \code{"BH"}, \code{"BY"}, and \code{"fdr"}. See \code{
\link[stats]{p.adjust}}.}

\item{sim.p}{Calculate \emph{p}-values via Monte Carlo simulation?}

\item{B}{Number of replicates or permutations to generate when computing
\emph{p}-values.}

\item{label}{Print association statistic over tiles?}

\item{diag}{Include principal diagonal of the concordance matrix? Only
advisable if \code{method = "MI"}.}

\item{title}{Optional plot title.}

\item{legend}{Legend position. Must be one of \code{"bottom"}, \code{"left"},
\code{"top"}, \code{"right"}, \code{"bottomright"}, \code{"bottomleft"},
\code{"topleft"}, or \code{"topright"}.}

\item{hover}{Show association statistic by hovering mouse over the
corresponding tile or circle? If \code{TRUE}, the plot is rendered in HTML
and will either open in your browser's graphic display or appear in the
RStudio viewer.}

\item{export}{Export concordance matrix? If \code{TRUE} and \code{alpha} is
non-\code{NULL}, then the \emph{p}-value matrix will also be returned.}
}
\value{
If \code{export = TRUE}, a list with up to two elements:
\itemize{
  \item The concordance matrix, computed via the chosen \code{method}.
  \item The matrix of \emph{p}-values (optionally adjusted), if \code{alpha}
  is non-\code{NULL}.
}
}
\description{
This function plots the pairwise concordance between categorical features.
}
\details{
Concordance plots visualize associations between categorical features. They
are useful when evaluating the dependencies between clinical factors and/or
patient clusters.

When \code{method = "fisher"}, concordance is measured by the negative
logarithm of the test's \emph{p}-value. If \code{sim.p = FALSE}, then the
function will attempt to calculate an exact \emph{p}-value. If this cannot be
executed in the available workspace, or if \code{sim.p = TRUE}, then \emph{
p}-values are estimated via Monte Carlo simulation with \code{B} replicates.

When \code{method = "chisq"}, concordance is measured by the Pearson
chi-squared statistic. This test is based on several assumptions that may not
be met in practice (see \href{http://bit.ly/2te2u5h}{Wikipedia} for a quick
overview). When one or several of these assumptions are violated, more
accurate \emph{p}-values can be estimated via Monte Carlo simulation with
\code{B} replicates.

When \code{method = "MI"}, concordance is measured by the mutual information
statistic. If \code{alpha} is non-\code{NULL}, then \emph{p}-values are
estimated via permutation testing with \code{B} permutations.
}
\examples{
df <- data.frame(A = sample.int(2, 20, replace = TRUE),
                 B = sample.int(3, 20, replace = TRUE),
                 C = sample.int(4, 20, replace = TRUE))
plot_concordance(df)

}
