% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_corr.R
\name{plot_corr}
\alias{plot_corr}
\title{Correlation Plot}
\usage{
plot_corr(
  dat,
  method = "pearson",
  use = "everything",
  alpha = NULL,
  p_adj = NULL,
  geom = "tile",
  label = FALSE,
  diag = FALSE,
  title = "Correlation Plot",
  legend = "right",
  hover = FALSE,
  export = FALSE
)
}
\arguments{
\item{dat}{A sample by feature data frame or matrix, e.g. of clinical
variables. Non-numeric features are dropped with a warning.}

\item{method}{String specifying which correlation coefficient to compute.
Must be one of \code{"pearson"}, \code{"kendall"}, or \code{"spearman"}.
See \code{\link[stats]{cor}}.}

\item{use}{Optional character string giving a method for computing
covariances in the presence of missing values. Must be one of \code{
"everything"}, \code{"all.obs"}, \code{"complete.obs"}, \code{
"na.or.complete"}, or \code{"pairwise.complete.obs"}.}

\item{alpha}{Optional significance threshold to impose on correlations.
Those with \emph{p}-values (optionally adjusted) less than or equal to
\code{alpha} are outlined in black.}

\item{p_adj}{Optional \emph{p}-value adjustment for multiple testing. Options
include \code{"holm"}, \code{"hochberg"}, \code{"hommel"}, \code{
"bonferroni"}, \code{"BH"}, \code{"BY"}, and \code{"fdr"}. See \code{
\link[stats]{p.adjust}}.}

\item{geom}{String specifying whether to visualize correlation coefficients
as \code{"tile"} or \code{"circle"}.}

\item{label}{Print correlation coefficient over \code{geom}?}

\item{diag}{Include principal diagonal of the correlation matrix?}

\item{title}{Optional plot title.}

\item{legend}{Legend position. Must be one of \code{"bottom"}, \code{"left"},
\code{"top"}, \code{"right"}, \code{"bottomright"}, \code{"bottomleft"},
\code{"topleft"}, or \code{"topright"}.}

\item{hover}{Show correlation coefficient by hovering mouse over the
corresponding tile or circle? If \code{TRUE}, the plot is rendered in HTML
and will either open in your browser's graphic display or appear in the
RStudio viewer.}

\item{export}{Export correlation matrix? If \code{TRUE} and \code{alpha} is
non-\code{NULL}, then the \emph{p}-value matrix will also be returned.}
}
\value{
If \code{export = TRUE}, a list with up to two elements:
\itemize{
  \item The concordance matrix, computed via the chosen \code{method}.
  \item The matrix of \emph{p}-values (optionally adjusted), if \code{alpha}
  is non-\code{NULL}.
}
}
\description{
This function plots pairwise correlations between variables.
}
\details{
Correlation plots visualize the associations between numeric features. They
are a valuable tool in exploratory data analysis for biological experiments,
where they may help identify dependencies among clinical covariates, leading
to better omic models.
}
\examples{
mat <- matrix(rnorm(100), 10, 10)
plot_corr(mat)

}
