% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_density.R
\name{plot_density}
\alias{plot_density}
\title{Create density plots by sample}
\usage{
plot_density(dat, group = NULL, type = NULL, xlab = NULL, main = NULL,
  legend = "outside", hover = FALSE)
}
\arguments{
\item{dat}{Omic data matrix with rows corresponding to probes and columns
to samples.}

\item{group}{Optional factor or character vector of length equal to sample size.
Levels are used to color density curves. If supplied, legend title defaults to
"Group". Override this feature by passing a named list instead.}

\item{type}{Optional string specifying omic data type. Currently supports
\code{"microarray", "RNA-seq",} or \code{"methylation"}.}

\item{xlab}{Label for x-axis. If left \code{NULL}, this defaults to
log expression for \code{type = "microarray"}, log CPM for
\code{type = "RNA-seq"}, or Beta for \code{type = "methylation"}. At least
one of \code{type} or \code{xlab} must be specified.}

\item{main}{Optional plot title.}

\item{legend}{Legend position. Must be one of \code{"outside",
"bottomleft", "bottomright", "topleft",} or \code{"topright"}.}

\item{hover}{Show sample name by hovering mouse over data point? If \code{TRUE},
the plot is rendered in HTML and will either open in your browser's graphic
display or appear in the RStudio viewer.}
}
\description{
Create density plots by sample
}
\details{
This function displays each sample's omic data distribution as a density curve.
It is especially helpful when contrasting pre- and post-normalization matrices.
It may additionally be used to inspect for batch effects or other associations
with phenotypic features by using the \code{group} argument.
}
\examples{
mat <- matrix(rnorm(5000), nrow = 1000, ncol = 5)
plot_density(mat, type = "microarray")

library(DESeq2)
mat <- cbind(matrix(rnbinom(5000, mu = 4, size = 1), nrow = 1000, ncol = 5),
             matrix(rnbinom(5000, mu = 4, size = 10), nrow = 1000, ncol = 5))
mat <- rlog(mat)
batch <- rep(c("A", "B"), each = 5)
plot_density(mat, group = batch, type = "RNA-seq",
             main = "rlog Transformed Counts")

}

