% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_density.R
\name{plot_density}
\alias{plot_density}
\title{Density Plots by Sample}
\usage{
plot_density(dat, group = NULL, xlab = NULL, title = NULL,
  legend = "outside", hover = FALSE)
}
\arguments{
\item{dat}{Omic data matrix or matrix-like object with rows corresponding to
probes and columns to samples.}

\item{group}{Optional character or factor vector of length equal to sample size.
Numeric or logical vectors will be silently coerced to factor. Levels are used
to color density curves. If supplied, legend title defaults to "Group". Override
this feature by passing a named list instead.}

\item{xlab}{Optional label for x-axis.}

\item{title}{Optional plot title.}

\item{legend}{Legend position. Must be one of \code{"outside", "bottomleft",
"bottomright", "topleft",} or \code{"topright"}.}

\item{hover}{Show sample name by hovering mouse over data point? If \code{TRUE},
the plot is rendered in HTML and will either open in your browser's graphic
display or appear in the RStudio viewer.}
}
\description{
This function displays each sample's omic data distribution as a density curve.
}
\details{
Density curves are an intuitive way to visualize an omic data distribution. They
are especially helpful when contrasting pre- and post-normalization matrices.
\code{plot_density} may additionally be used to inspect for batch effects
or associations with phenotypic factors by using the \code{group} argument.
}
\examples{
# Density plots by sample
mat <- matrix(rnorm(500 * 10), nrow = 500, ncol = 10)
plot_density(mat)

# Density plots by group
grp <- rep(c("A", "B"), each = 5)
plot_density(mat, group = grp)

}
\seealso{
\code{\link[limma]{plotDensities}}
}
