% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dispersion.R
\name{plot_dispersion}
\alias{plot_dispersion}
\alias{plot_dispersion.DGEList}
\alias{plot_dispersion.DESeqDataSet}
\alias{plot_dispersion.default}
\title{Plot Dispersion Estimates}
\usage{
plot_dispersion(
  dat,
  trans = "log",
  title = "Mean-Dispersion Plot",
  legend = "right",
  ...
)

\method{plot_dispersion}{DGEList}(
  dat,
  design = NULL,
  trans = "log",
  size = NULL,
  alpha = NULL,
  title = "Mean-Dispersion Plot",
  legend = "right",
  hover = FALSE
)

\method{plot_dispersion}{DESeqDataSet}(
  dat,
  trans = "log",
  size = NULL,
  alpha = NULL,
  title = "Mean-Dispersion Plot",
  legend = "right",
  hover = FALSE
)

\method{plot_dispersion}{default}(
  dat,
  design = NULL,
  trans = "log",
  pipeline = NULL,
  size = NULL,
  alpha = NULL,
  title = "Mean-Dispersion Plot",
  legend = "right",
  hover = FALSE
)
}
\arguments{
\item{dat}{A \code{\link[edgeR]{DGEList}} object or \code{\link[DESeq2]{
DESeqDataSet}}. If normalization factors or dispersions have not already
been estimated for \code{dat}, they will be internally computed using the
appropriate \pkg{edgeR} or \pkg{DESeq2} functions. Alternatively, \code{
dat} may be a raw count matrix, in which case the \code{pipeline} argument
must specify which package to use for estimating normalization factors and
genewise dispersions. A design matrix may be required to calculate adjusted
profile log-likelihoods. See Details.}

\item{trans}{Data transformation to be applied to genewise dispersion
estimates. Must be one of \code{"log"} or \code{"sqrt"}.}

\item{title}{Optional plot title.}

\item{legend}{Legend position. Must be one of \code{"bottom"}, \code{"left"},
\code{"top"}, \code{"right"}, \code{"bottomright"}, \code{"bottomleft"},
\code{"topleft"}, or \code{"topright"}.}

\item{design}{Optional design matrix with rows corresponding to samples and
columns to model coefficients. This will be extracted from \code{dat} if 
available and need not be set explicitly in the call. If provided, however, 
\code{design} will override the relevant slot of \code{dat}.}

\item{size}{Point size.}

\item{alpha}{Point transparency.}

\item{hover}{Show probe name by hovering mouse over data point? If \code{
TRUE}, the plot is rendered in HTML and will either open in your browser's
graphic display or appear in the RStudio viewer.}

\item{pipeline}{Which package should be used to estimate normalization
factors and genewise dispersions? Only relevant if \code{dat} is a raw
count matrix. Must be one of \code{"edgeR"} or \code{"DESeq2"}. Default
settings are applied at all steps. For greater control of internal
parameters, create the appropriate \code{DGEList} or \code{DESeqDataSet}
object with your preferred settings and pass it directly as \code{dat}.}
}
\description{
This function plots the mean-dispersion relationship for a given count
matrix.
}
\details{
Count data in omic studies are often presumed to follow a negative binomial
distribution, which may be uniquely identified by its mean and dispersion
parameters. For RNA-seq pipelines that rely on negative binomial generalized
linear models (GLMs), such as \code{edgeR} and \code{DESeq2}, estimating
genewise dispersions is therefore an essential step in the model fitting
process. Because there are rarely sufficient replicates to reliably infer
these values independently for each gene, both packages use empirical Bayes
methods to pool information across genes.

Details vary between the two pipelines, which is why outputs will differ for
\code{DGEList} objects and \code{DESeqDataSet}s. \code{edgeR} begins by
computing a common dispersion parameter for the entire dataset, rendered by
\code{plot_dispersion} as a blue horizontal line; then fits a trend line to
the maximized negative binomial likelihoods, represented by an orange curve;
and finally calculates posterior estimates, depicted by black points, using a
weighted empirical Bayes likelihood procedure. Be aware that likelihood
maximization methods for \code{DGEList} objects vary depending on whether or
not a model matrix is supplied. See \code{\link[edgeR]{estimateDisp}} for
more details. A thorough explication of the statistical theory behind this
pipeline can be found in the original papers by the packages authors:
Robinson & Smyth (2007), Robinson & Smyth (2008), and McCarthy et al. (2012).

\code{DESeq2} also fits a trend line through likelihood estimates of genewise
dispersions, depicted by orange points in the \code{plot_dispersion} output.
Posterior values are calculated following regression toward a log-normal
prior with mean equal to the predicted value from the trended fit and
variance equal to the difference between the observed variance of the log
dispersion estimates and the expected sampling variance. These maximum a
posteriori values are colored blue, while outliers, defined as genes with log
dispersion values more than two median absolute deviations away from the
trend line, are colored red. See \code{\link[DESeq2]{estimateDispersions}}
for more details. For more comprehensive statistical background, see the
original DESeq paper (Anders & Huber, 2010) and the DESeq2 paper (Love et
al., 2014).

\code{plot_dispersion} effectively combines 
\code{edgeR::\link[edgeR]{plotBCV}} and 
\code{DESeq2::\link[DESeq2]{plotDispEsts}} into a single function that can 
take either a \code{DGEList} or a \code{DESeqDataSet} as its argument and 
return the matching figure. By default, dispersions are plotted under log10 
transform. They may also be displayed under square root transform, in which 
case the y-axis can be interpreted as the biological coefficient of variation 
(McCarthy et al., 2012).
}
\examples{
library(DESeq2)
dds <- makeExampleDESeqDataSet()
plot_dispersion(dds)

# Plot the same data using the edgeR pipeline and sqrt transform
plot_dispersion(counts(dds), design = model.matrix(~ condition, colData(dds)),
                pipeline = "edgeR", trans = "sqrt")

}
\references{
Anders, S. & Huber, W. (2010).
\href{http://bit.ly/2sJrnAU}{Differential expression analysis for sequence
count data}. \emph{Genome Biology}, 11:R106.

McCarthy, D.J., Chen, Y., & Smyth, G.K. (2012).
\href{http://dx.doi.org/10.1093/nar/gks042}{Differential expression analysis
of multifactor RNA-Seq experiments with respect to biological variation}.
\emph{Nucleic Acids Res., 40}(10): 4288-4297.

Love, M., Huber, W. & Anders, S. (2014).
\href{http://bit.ly/2pytbeL}{Moderated estimation of fold change and
dispersion for RNA-seq data with DESeq2}. \emph{Genome Biology, 15}(12): 550.

Robinson, M.D. and Smyth, G.K. (2008).
\href{http://biostatistics.oxfordjournals.org/content/9/2/321}{Small-sample
estimation of negative binomial dispersion, with applications to SAGE data}.
\emph{Biostatistics, 9}(2):321-332.

Robinson, M.D. & Smyth, G.K. (2007).
\href{http://bioinformatics.oxfordjournals.org/content/23/21/2881}{Moderated
statistical tests for assessing differences in tag abundance}.
\emph{Bioinformatics, 23}(21): 2881-2887.
}
\seealso{
\code{\link[DESeq2]{plotDispEsts}, \link[edgeR]{plotBCV},
\link[DESeq2]{estimateDispersions}, \link[edgeR]{estimateDisp}}
}
