% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_drivers.R
\name{plot_drivers}
\alias{plot_drivers}
\title{Plot Drivers of Omic Variation}
\usage{
plot_drivers(
  dat,
  clin,
  parametric = TRUE,
  block = NULL,
  unblock = NULL,
  kernel = NULL,
  kpar = NULL,
  top = NULL,
  n.pc = 5L,
  label = FALSE,
  alpha = NULL,
  p.adj = NULL,
  title = "Variation By Feature",
  legend = "right",
  hover = FALSE
)
}
\arguments{
\item{dat}{Omic data matrix or matrix-like object with rows corresponding to
probes and columns to samples. It is strongly recommended that data be
filtered and normalized prior to plotting. Raw counts stored in 
\code{\link[edgeR]{DGEList}} or \code{\link[DESeq2]{DESeqDataSet}} objects 
are automatically extracted and transformed to the log2-CPM scale, with a
warning.}

\item{clin}{Data frame or matrix with rows correponding to samples and
columns to technical and/or biological features to test for associations
with omic data.}

\item{parametric}{Compute \emph{p}-values using parametric association tests?
If \code{FALSE}, rank-based alternatives are used instead. See Details.}

\item{block}{String specifying the name of the column in which to find the
blocking variable, should one be accounted for. See Details.}

\item{unblock}{Column name(s) of one or more features for which the 
\code{block} covariate should not be applied, if one was supplied. See 
Details.}

\item{kernel}{The kernel generating function, if using KPCA. Options include
\code{"rbfdot"}, \code{"polydot"}, \code{"tanhdot"}, \code{"vanilladot"}, 
\code{"laplacedot"}, \code{"besseldot"}, \code{"anovadot"}, and 
\code{"splinedot"}. To run normal PCA, set to \code{NULL}. See Details.}

\item{kpar}{A named list of arguments setting parameters for the kernel
function. Only relevant if \code{kernel} is not \code{NULL}. See Details.}

\item{top}{Optional number (if > 1) or proportion (if < 1) of most variable
probes to be used for PCA.}

\item{n.pc}{Number of principal components to include in the figure.}

\item{label}{Print association statistics over tiles?}

\item{alpha}{Optional significance threshold to impose on associations. 
Those with \emph{p}-values (optionally adjusted) less than or equal to 
\code{alpha} are outlined in black.}

\item{p.adj}{Optional \emph{p}-value adjustment for multiple testing. Options
include \code{"holm"}, \code{"hochberg"}, \code{"hommel"}, \code{
"bonferroni"}, \code{"BH"}, \code{"BY"}, and \code{"fdr"}. See \code{
\link[stats]{p.adjust}}.}

\item{title}{Optional plot title.}

\item{legend}{Legend position. Must be one of \code{"bottom"}, \code{"left"},
\code{"top"}, \code{"right"}, \code{"bottomright"}, \code{"bottomleft"},
\code{"topleft"}, or \code{"topright"}.}

\item{hover}{Show \emph{p}-values by hovering mouse over tiles? If 
\code{TRUE}, the plot is rendered in HTML and will either open in your 
browser's graphic display or appear in the RStudio viewer.}
}
\description{
This function visualizes the strength of associations between the principal
components of an omic data matrix and a set of biological and/or technical
features.
}
\details{
Strength of association is measured by -log \emph{p}-values, optionally
adjusted for multiple testing. When \code{parametric = TRUE}, significance
is computed from Pearson correlation tests (for continuous features) or 
ANOVA \emph{F}-tests (for categorical features). When 
\code{parametric = FALSE}, significance is computed from Spearman correlation
tests (for continuous features) or Kruskal-Wallis tests (for categorical
features). 

An optional blocking variable may be provided if samples violate the
assumption of independence, e.g. for studies in which subjects are observed
at multiple time points. If a blocking variable is identified, it will be
regressed out prior to testing for all variables except those explicitly
exempted by the \code{unblock} argument. 

When supplying a blocking variable, be careful to consider potential
confounding effects. For instance, features like sex and age are usually
nested within subject, while subject may be nested within other variables
like batch or treatment group. The \code{block} and \code{unblock} arguments
are designed to help parse out these relationships.

If \code{kernel} is non-\code{NULL}, then KPCA is used instead of PCA. See
\code{\link{plot_kpca}} for more info. Details on kernel functions and their
input parameters can be found in \code{kernlab::\link[kernlab]{dots}}.
}
\examples{
library(SummarizedExperiment)
library(edgeR)
library(dplyr)
data(airway)
cnts <- assay(airway)
keep <- rowSums(cpm(cnts) > 1) >= 4
mat <- cpm(cnts[keep, ], log = TRUE)
clin <- colData(airway) \%>\%
  as_tibble(.) \%>\%
  select(Run, cell, dex)
plot_drivers(mat, clin)

}
\seealso{
\code{\link{plot_pca}}, \code{\link{plot_kpca}}
}
