% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Omic Heatmap}
\usage{
plot_heatmap(dat, group = NULL, covar = NULL, dist = "pearson",
  p = 2L, top = NULL, filter_method = "pairwise", center = FALSE,
  hclustfun = "average", pal_group = "npg", pal_covar = "Blues",
  pal_tiles = "RdBu", title = "Omic Heatmap")
}
\arguments{
\item{dat}{Omic data matrix or matrix-like object with rows corresponding to
probes and columns to samples.}

\item{group}{Optional character or factor vector of length equal to sample
size. Alternatively, a data frame or list of such vectors, optionally
named.}

\item{covar}{Optional continuous covariate of length equal to sample size.
Alternatively, a data frame or list of such vectors, optionally named.}

\item{dist}{Distance measure to be used. Supports all methods available in
\code{\link[stats]{dist}}, \code{Rfast::\link[Rfast]{Dist}}, and \code{
\link[vegan]{vegdist}}, as well as those implemented in the \code{bioDist} 
package. See Details.}

\item{p}{Power of the Minkowski distance.}

\item{top}{Optional number (if > 1) or proportion (if < 1) of top probes to
be used for t-SNE.}

\item{filter_method}{String specifying whether to apply a \code{"pairwise"}
or \code{"common"} filter if \code{top} is non-\code{NULL}.}

\item{center}{Center each probe prior to computing distances?}

\item{hclustfun}{The agglomeration method to be used for hierarchical
clustering. Supports all methods available in \code{\link[stats]{hclust}}.}

\item{pal_group}{String specifying the color palette to use if \code{group}
is non-\code{NULL}, or a vector of such strings with length equal to the
number of vectors passed to \code{group}. Options include \code{"ggplot"}, 
all qualitative color schemes available in \code{\href{
https://bit.ly/2ipuEjn}{RColorBrewer}}, and the complete collection of 
\code{\href{http://bit.ly/2bxnuGB}{ggsci}} palettes. Alternatively, a 
character vector of colors with length equal to the cumulative number of 
levels in \code{group}.}

\item{pal_covar}{String specifying the color palette to use if \code{covar}
is non-\code{NULL}, or a vector of such strings with length equal to the
number of vectors passed to \code{covar}. Options include the complete 
collection of \code{\href{https://bit.ly/2n7D6tF}{viridis}} palettes, as 
well as all sequential color schemes available in \code{RColorBrewer}. 
Alternatively, a character vector of colors representing a smooth gradient, 
or a list of such vectors with length equal to the number of continuous 
variables to visualize.}

\item{pal_tiles}{String specifying the color palette to use for heatmap
tiles. Options include the complete collection of \code{\href{
https://bit.ly/2n7D6tF}{viridis}} palettes, as well all diverging color 
schemes available in \code{\href{https://bit.ly/2ipuEjn}{RColorBrewer}}. 
Alternatively, a character vector of at least two colors.}

\item{title}{Optional plot title.}
}
\description{
This function visualizes a probe by sample matrix as a heatmap.
}
\details{
Heatmaps are a common and intuitive way to display the values of an omic data
matrix, especially after top probes have been selected for closer
investigation. Hierarchical clustering dendrograms cluster both the rows and
the columns, revealing latent structure in the data. Annotation tracks atop
the figure may be used to investigate associations with phenotypic features.

Available distance measures include: \code{"euclidean"}, \code{"maximum"},
\code{"manhattan"}, \code{"canberra"}, \code{"minkowski"}, \code{"cosine"},
\code{"pearson"}, \code{"kendall"}, \code{"spearman"}, \code{"bray"}, \code{
"kulczynski"}, \code{"jaccard"}, \code{"gower"}, \code{"altGower"}, \code{
"morisita"}, \code{"horn"}, \code{"mountford"}, \code{"raup"}, \code{
"binomial"}, \code{"chao"}, \code{"cao"}, \code{"mahalanobis"}, \code{"MI"},
or \code{"KLD"}. Some distance measures are unsuitable for certain types of
data. See \code{\link{dist_mat}} for more details on these methods and links
to documentation on each.

The \code{top} argument optionally filters data using either probewise
variance (if \code{filter_method = "common"}) or the leading fold change
method of Smyth et al. (if \code{filter_method = "pairwise"}). See \code{
\link{plot_mds}} for more details.
}
\examples{
mat <- matrix(rnorm(100 * 10), nrow = 100, ncol = 10)
grp <- rep(c("A", "B"), each = 5)
plot_heatmap(mat, group = grp)

}
