% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Omic Heatmap}
\usage{
plot_heatmap(dat, group = NULL, covar = NULL, dist = "pearson",
  hclustfun = "average", pal_group = "npg", pal_covar = "blues",
  pal_tiles = "RdBu", title = NULL)
}
\arguments{
\item{dat}{Omic data matrix or matrix-like object with rows corresponding to
probes and columns to samples.}

\item{group}{Optional character or factor vector of length equal to sample
size. Alternatively, a data frame or list of such vectors, optionally
named.}

\item{covar}{Optional continuous covariate of length equal to sample size.
Alternatively, a data frame or list of such vectors, optionally named.}

\item{dist}{Distance measure to be used. Currently supports any method
available in \code{\link[stats]{dist}} or \code{\link[stats]{cor}}.}

\item{hclustfun}{The agglomeration method to be used for hierarchical
clustering. Supports any method available in \code{\link[stats]{hclust}}.}

\item{pal_group}{String specifying the color palette to use if \code{group}
is non-\code{NULL}, or a vector of such strings with length equal to the
number of vectors passed to \code{group}. Options include \code{'ggplot'},
as well as the complete collection of \code{
\href{https://cran.r-project.org/web/packages/ggsci/vignettes/ggsci.html}{
ggsci}} palettes, which can be identified by name (e.g., \code{'npg'},
\code{'aaas'}, etc.). Alternatively, any character vector of colors with
length equal to the cumulative number of levels in \code{group}.}

\item{pal_covar}{String specifying the color palette to use if \code{covar}
is non-\code{NULL}, or a vector of such strings with length equal to the
number of vectors passed to \code{covar}. Options include \code{'blues'},
\code{'greens'}, \code{'purples'}, \code{'greys'}, \code{'oranges'}, and
\code{'reds'}. Alternatively, any character vector of colors representing
a smooth gradient, or a list of such vectors with length equal to the
number of continuous variables to visualize.}

\item{pal_tiles}{String specifying the color palette to use for heatmap
tiles. Options include \code{'RdBu'} for red to blue gradient, \code{
'GrRd'} for green to red gradient, and \code{'BuYl'} for blue to yellow
gradient. Alternatively, any user-supplied color palette is acceptable.}

\item{title}{Optional plot title.}
}
\description{
This function visualizes a probe by sample matrix as a heatmap.
}
\details{
Heatmaps are a common and intuitive way to display the values of an omic data
matrix, especially after top probes have been selected for closer
investigation. Hierarchical clustering dendrograms cluster both the rows and
the columns, revealing latent structure in the data. Annotation tracks atop
the figure may be used to investigate associations with phenotypic features.
}
\examples{
mat <- matrix(rnorm(100 * 10), nrow = 100, ncol = 10)
grp <- rep(c("A", "B"), each = 5)
plot_heatmap(mat, group = grp)

}
