% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Omic Heatmap}
\usage{
plot_heatmap(dat, anno = NULL, dist = "pearson", hclustfun = "average",
  col = "RdBu", title = NULL)
}
\arguments{
\item{dat}{Omic data matrix or matrix-like object with rows corresponding to
probes and columns to samples.}

\item{anno}{Optional character, factor, numeric, or logical vector of length
equal to sample size. Alternatively, a data frame or list of such vectors,
optionally named. Values are used to color one or several annotation tracks
atop the heatmap.}

\item{dist}{Distance measure to be used. Currently supports any method available
in \code{\link[stats]{dist}} or \code{\link[stats]{cor}}.}

\item{hclustfun}{The agglomeration method to be used for hierarchical clustering.
See \code{\link[stats]{hclust}} for available options.}

\item{col}{Color palette to use for heatmap tiles. Preset options include \code{
"RdBu"} for red to blue gradient, \code{"GrRd"} for green to red gradient, and
\code{"BuYl"} for blue to yellow gradient. Alternatively, any user-supplied
color palette is acceptable.}

\item{title}{Optional plot title.}
}
\description{
This function visualizes a probe by sample matrix as a heatmap.
}
\details{
Heatmaps are a common and intuitive way to display the values of an omic data
matrix, especially after top probes have been selected for closer investigation.
Hierarchical clustering dendrograms cluster both the rows and the columns,
revealing latent structure in the data. Annotation tracks atop the figure may be
used to investigate associations with phenotypic features.
}
\examples{
mat <- matrix(rnorm(100 * 10), nrow = 100, ncol = 10)
grp <- rep(c("A", "B"), each = 5)
plot_heatmap(mat, anno = grp)

}
