% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_md.R
\name{plot_md}
\alias{plot_md}
\title{Mean-Difference Plot}
\usage{
plot_md(dat, fdr = 0.05, ptsize = 0.25, main = NULL, legend = "outside",
  hover = FALSE, probes = NULL)
}
\arguments{
\item{dat}{Data frame or matrix representing the results of a test for
differential expression or methylation, such as the output of a call to
\code{\link[limma]{topTable}}, \code{\link[edgeR]{topTags}}, or
\code{\link[DESeq2]{results}}. Alternatively, any object with columns for log
fold changes, probewise means, and FDR.}

\item{fdr}{Threshold for declaring a probe differentially expressed or methylated.}

\item{ptsize}{Size of data points in the plot.}

\item{main}{Optional plot title.}

\item{legend}{Legend position. Must be one of \code{"outside",
"bottomleft", "bottomright", "topleft",} or \code{"topright"}.}

\item{hover}{Show probe name by hovering mouse over data point? If \code{TRUE},
the plot is rendered in HTML and will either open in your browser's graphic
display or appear in the RStudio viewer.}

\item{probes}{String specifying the name of the column in which to find the probe
identifiers, assuming they are not \code{rownames(dat)}. Only relevant if
\code{hover = TRUE}.}
}
\description{
This function plots probewise means against the log2 fold changes of a
differential expression or methylation test.
}
\details{
MD plots (also known as "Bland-Altman plots" or "MA plots") visualize the
relationship between a probe's mean value and its log2 fold change for a given
test of differential expression/methylation. Points are colored to distinguish
between those that do and do not meet a user-defined FDR threshold. These
figures help to evaluate the symmetry, magnitude, and significance of effects
for a given experiment.
}
\examples{
library(limma)
DE_genes <- cbind(matrix(rnorm(50 * 5, mean = 5), nrow = 50, ncol = 5),
                  matrix(rnorm(50 * 5), nrow = 50, ncol = 5))
mat <- rbind(DE_genes, matrix(rnorm(4950 * 10), nrow = 4950, ncol = 10))
treat <- rep(c("A", "B"), each = 5)
des <- model.matrix(~ treat)
fit <- eBayes(lmFit(mat, des))
top <- topTable(fit, number = Inf)
plot_md(top)

}
