% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_md.R
\name{plot_md}
\alias{plot_md}
\title{Create MD plot of \eqn{log_2} fold changes against mean expression or methylation}
\usage{
plot_md(dat, fdr = 0.05, ptsize = 0.25, main = NULL, legend = "outside",
  hover = FALSE, knitr = FALSE)
}
\arguments{
\item{dat}{A count}

\item{fdr}{Threshold for declaring a probe differentially expressed or methylated.}

\item{ptsize}{Size of data points in the plot.}

\item{main}{Optional plot title.}

\item{legend}{Legend position. Must be one of \code{"outside",
"bottomleft", "bottomright", "topleft",} or \code{"topright"}.}

\item{hover}{Show probe name by hovering mouse over data point? If \code{TRUE},
the plot is rendered in HTML and will either open in your browser's graphic
display or appear in the RStudio viewer. The plot can also be embedded in an
HTML doc using Rmarkdown so long as \code{knitr = TRUE} and the code chunk
option \code{plotly} is also set to \code{TRUE}.}

\item{knitr}{Set this to \code{TRUE} if you want to embed a plotly object (viz.,
the \code{plot_md} output when \code{hover = TRUE}) in an HTML doc. Make
sure to set \code{plotly = TRUE} in the corresponding code chunk options.}

\item{res}{Data frame or matrix representing the results of a test for
differential expression or methylation, such as the output of a call to
limma::topTable, edgeR::topTags, or DESeq2::results. Alternatively, any
object with columns for log fold changes, and FDR.}

\item{type}{String specifying data type. Must be one of either
\code{"microarray"}, \code{"RNA-seq"}, or \code{"methylation"}.}
}
\description{
Create MD plot of \eqn{log_2} fold changes against mean expression or methylation
}
\details{
This function displays the results of a differential expression or methylation
test as an MD plot, with colors distinguishing probes that meet the
user-defined FDR threshold. These figures can be used to evaluate the symmetry,
magnitude, and significance of fold changes for a given experiment.
}
\examples{
library(dplyr)
df <- data_frame(logFC   = c(rnorm(50, 0, 10), rnorm(4950)),
                 AvgExpr = rowMeans(matrix(rnorm(5000), nrow = 5000, ncol = 10)),
                 p.value = pnorm(-abs(logFC)),
                 FDR     = p.adjust(p.value, method = "fdr"))
plot_md(df)

library(limma)
DE_genes <- cbind(matrix(rnorm(250, 5, 1), nrow = 50, ncol = 5),
                  matrix(rnorm(250), nrow = 50, ncol = 5))
mat <- rbind(DE_genes, matrix(rnorm(45500), nrow = 4550, ncol = 10))
treat <- rep(c("A", "B"), each = 5)
des <- model.matrix(~ treat)
fit <- eBayes(lmFit(mat, des))
top <- topTable(fit, number = Inf)
plot_md(top)

}

