% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_md.R
\name{plot_md}
\alias{plot_md}
\alias{plot_md.DGEList}
\alias{plot_md.DESeqDataSet}
\alias{plot_md.DESeqTransform}
\alias{plot_md.DESeqResults}
\alias{plot_md.TopTags}
\alias{plot_md.data.frame}
\alias{plot_md.default}
\title{Mean-Difference Plot}
\usage{
plot_md(dat, title = NULL, legend = "right", ...)

\method{plot_md}{DGEList}(dat, design = NULL, sample = 1L,
  ctrls = NULL, lfc = NULL, title = NULL, xlab = NULL,
  legend = "right", hover = FALSE)

\method{plot_md}{DESeqDataSet}(dat, sample = 1L, ctrls = NULL,
  lfc = NULL, title = NULL, xlab = NULL, legend = "right",
  hover = FALSE)

\method{plot_md}{DESeqTransform}(dat, sample = 1L, ctrls = NULL,
  lfc = NULL, title = NULL, xlab = NULL, legend = "right",
  hover = FALSE)

\method{plot_md}{DESeqResults}(dat, fdr = 0.05, lfc = NULL,
  title = NULL, xlab = NULL, legend = "right", hover = FALSE)

\method{plot_md}{TopTags}(dat, fdr = 0.05, lfc = NULL, title = NULL,
  xlab = NULL, legend = "right", hover = FALSE)

\method{plot_md}{data.frame}(dat, probes = NULL, fdr = 0.05,
  lfc = NULL, title = NULL, xlab = NULL, legend = "right",
  hover = FALSE)

\method{plot_md}{default}(dat, sample = 1L, ctrls = NULL, lfc = NULL,
  title = NULL, xlab = NULL, legend = "right", hover = FALSE)
}
\arguments{
\item{dat}{Either a data frame representing the results of a test for
differential expression, or a probe by sample omic data matrix. The former
will render a study-wide MD plot, the latter a between-sample MD plot.
Suitable objects from familiar packages are also acceptable. See Details.}

\item{title}{Optional plot title.}

\item{legend}{Legend position. Must be one of \code{"bottom"}, \code{"left"},
\code{"top"}, \code{"right"}, \code{"bottomright"}, \code{"bottomleft"},
\code{"topleft"}, or \code{"topright"}.}

\item{design}{Optional design matrix with rows corresponding to samples and
columns to coefficients to be estimated. Only relevant for \code{
\link[edgeR]{DGEList}} objects.}

\item{sample}{Column number or name specifying which sample in \code{dat} to
compare with the others. Only relevant for between-sample MD plots.}

\item{ctrls}{Optional vector of length equal to \code{nrow(dat)} indicating
the control status of each probe. Only relevant for between-sample MD
plots.}

\item{lfc}{Optional effect size threshold for declaring a probe
differentially expressed. Only relevant for study-wide MD plots.}

\item{xlab}{Optional label for x-axis.}

\item{hover}{Show probe name by hovering mouse over data point? If \code{
TRUE}, the plot is rendered in HTML and will either open in your browser's
graphic display or appear in the RStudio viewer. Probe names are extracted
from \code{dat}.}

\item{fdr}{Optional significance threshold for declaring a probe
differentially expressed. Only relevant for study-wide MD plots.}

\item{probes}{Optional column number or name specifying where probe names are
stored, presuming they are not stored in \code{rownames(dat)}.}
}
\description{
This function plots probewise means vs. log2 fold changes for a test of
differential expression or between-sample comparison.
}
\details{
MD plots (also known as "Bland-Altman plots" or "MA plots") visualize the
relationship between a probe's mean value and its log2 fold change versus
some relevant reference group. These figures help to evaluate the symmetry,
magnitude, and significance of differential effects across the full omic
range.

If \code{dat} summarizes the results of a test for differential expression,
then each point's x-coordinate correponds to its average expression across
all samples, while y-coordinates represent the log2 fold change for the given
contrast. Points are colored to distinguish between those that do and do not
meet a user-defined FDR threshold. \code{plot_md} accepts output from
\code{limma::\link[limma]{topTable}}, \code{edgeR::\link[edgeR]{topTags}}, or
\code{DESeq2::\link[DESeq2]{results}}. Alternatively, any object with columns
for log fold changes, probewise means, and FDR is acceptable.

If \code{dat} is probe by sample matrix or matrix-like object, then \code{
sample} must be specified. An artificial array is created by averaging
probewise values for all other samples in the data. The figure will then
represent the mean vs. the difference of expression values for the specified
sample vs. the artificial array. Acceptable inputs for between-sample MD
plots include all \code{limma} expression set objects, as well as \code{
\link[edgeR]{DGEList}}, \code{\link[DESeq2]{DESeqDataSet}}, and \code{
\link[DESeq2]{DESeqTransform}} objects.
}
\examples{
library(DESeq2)
dds <- makeExampleDESeqDataSet()

# Between-sample MD plot
rld <- rlog(dds)
plot_md(rld)

# Study-wide MD plot
dds <- DESeq(dds)
res <- results(dds)
plot_md(res)

}
\references{
Bolstad, B.M., Irizarry, R.A., Åstrand, M. & Speed, T.P. (2003).
\href{https://www.ncbi.nlm.nih.gov/pubmed/12538238}{A comparison of
normalization methods for high density oligonucleotide array data based on
variance and bias}. \emph{Bioinformatics}, \emph{19}(2): 185–193.

Dudoit, S., Yang, Y.H., Callow, M.J. & Speed, T.P. (2002).
\href{https://www.jstor.org/stable/24307038?seq=1#page_scan_tab_contents}{
Statistical methods for identifying differentially expressed genes in
replicated cDNA microarray experiments}. \emph{Stat. Sin.}, \strong{12},
111–140.

Martin, B.J. & Altman, D.G. (1986).
\href{http://www.sciencedirect.com/science/article/pii/S0140673686908378}{
Statistical methods for assessing agreement between two methods of clinical
measurement}. \emph{The Lancet}, \emph{327}(8476): 307–310.

Ritchie, M.E., Phipson, B., Wu, D., Hu, Y., Law, C.W., Shi, W., & Smyth, G.K.
(2015).
\href{https://www.ncbi.nlm.nih.gov/pubmed/25605792}{limma powers differential
expression analyses for RNA-sequencing and microarray studies}. \emph{Nucleic
Acids Res.}, \emph{43}(7): e47.
}
\seealso{
\code{\link[limma]{plotMD}}, \code{\link[DESeq2]{plotMA}},
\code{\link[Glimma]{glMDPlot}}
}
