% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_md.R
\name{plot_md}
\alias{plot_md}
\title{Mean-Difference Plot}
\usage{
plot_md(dat, fdr = 0.05, lfc = NULL, sample = NULL, ctrls = NULL,
  title = NULL, legend = "outside", hover = FALSE)
}
\arguments{
\item{dat}{Either a data frame representing the results of a test for differential
expression, or an omic data matrix with rows corresponding to probes and columns
to samples. The former will render a study-wide MD plot, the latter a between
between-sample MD plot. See Details.}

\item{fdr}{Significance threshold for declaring a probe differentially expressed. Only
relevant for study-wide MD plots.}

\item{lfc}{Optional effect size threshold for declaring a probe differentially
expressed. Only relevant for study-wide MD plots.}

\item{sample}{Column number or name specifying which sample in \code{dat} to
compare with the others. Only relevant for between-sample MD plots.}

\item{ctrls}{Optional vector of length equal to \code{nrow(dat)} indicating the
control status of each probe. Only relevant for between-sample MD plots.}

\item{title}{Optional plot title.}

\item{legend}{Legend position. Must be one of \code{"outside", "bottomleft",
"bottomright", "topleft",} or \code{"topright"}.}

\item{hover}{Show probe name by hovering mouse over data point? If \code{TRUE},
the plot is rendered in HTML and will either open in your browser's graphic
display or appear in the RStudio viewer. Probe names are extracted from
\code{dat}.}
}
\description{
This function plots probewise means vs. log2 fold changes for a test of
differential expression or between-sample comparison.
}
\details{
MD plots (also known as "Bland-Altman plots" or "MA plots") visualize the
relationship between a probe's mean value and its log2 fold change versus some
relevant reference group. These figures help to evaluate the symmetry, magnitude,
and significance of differential effects across the full omic range.

If \code{dat} is a data frame summarizing the results of a test for differential
expression, then each point's x-coordinate correponds to its average expression
across all samples, while y-coordinates represent the log2 fold change for the
given contrast. Points are colored to distinguish between those that do and do not
meet a user-defined FDR threshold. \code{plot_md} accepts output from \code{
limma::\link[limma]{topTable}}, \code{DESeq2::\link[DESeq2]{results}}, or \code{
edgeR::\link[edgeR]{topTags}}. Alternatively, any object with columns for log fold
changes, probewise means, and FDR is acceptable.

If \code{dat} is probe by sample matrix or matrix-like object, then \code{sample}
must be specified. An artificial array is created by averaging probewise values for
all other samples in the data. The figure will then represent the mean vs. the
difference of expression values for the specified sample vs. the artificial array.
}
\examples{
library(limma)
DE_genes <- cbind(matrix(rnorm(50 * 5, mean = 5), nrow = 50, ncol = 5),
                  matrix(rnorm(50 * 5), nrow = 50, ncol = 5))
mat <- rbind(DE_genes, matrix(rnorm(4950 * 10), nrow = 4950, ncol = 10))
treat <- rep(c("A", "B"), each = 5)
des <- model.matrix(~ treat)
fit <- eBayes(lmFit(mat, des))
top <- topTable(fit, number = Inf)
plot_md(top)

}
\seealso{
\code{\link[limma]{plotMD}}, \code{\link[DESeq2]{plotMA}},
\code{\link[Glimma]{glMDPlot}}
}
