% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mds.R
\name{plot_mds}
\alias{plot_mds}
\title{MDS Plot}
\usage{
plot_mds(dat, group = NULL, covar = NULL, metric = TRUE,
  dist = "euclidean", p = 2L, top = 500L,
  filter_method = "pairwise", pcs = c(1L, 2L), label = FALSE,
  pal_group = "npg", pal_covar = "Blues", size = NULL,
  alpha = NULL, title = "MDS", legend = "right", hover = FALSE,
  D3 = FALSE)
}
\arguments{
\item{dat}{Omic data matrix or matrix-like object with rows corresponding to
probes and columns to samples. It is strongly recommended that data be
filtered and normalized prior to plotting. Raw counts stored in \code{
\link[edgeR]{DGEList}} or \code{\link[DESeq2]{DESeqDataSet}} objects are
automatically extracted and transformed to the log2-CPM scale, with a
warning. Alternatively, an object of class \code{dist} which can be
directly input to the MDS algorithm.}

\item{group}{Optional character or factor vector of length equal to sample
size, or up to two such vectors organized into a list or data frame. Supply
legend title(s) by passing a named list or data frame.}

\item{covar}{Optional continuous covariate. If non-\code{NULL}, then plot can
render at most one \code{group} variable. Supply legend title by passing
a named list or data frame.}

\item{metric}{Logical. Perform classical (i.e. metric) MDS or nonmetric MDS?
See Details.}

\item{dist}{Distance measure to be used. Supports all methods available in
\code{\link[stats]{dist}}, \code{Rfast::\link[Rfast]{Dist}}, and \code{
\link[vegan]{vegdist}}, as well as those implemented in the \code{bioDist} 
package. See Details.}

\item{p}{Power of the Minkowski distance.}

\item{top}{Optional number (if > 1) or proportion (if < 1) of top probes to
be used for MDS.}

\item{filter_method}{String specifying whether to apply a \code{"pairwise"}
or \code{"common"} filter if \code{top} is non-\code{NULL}. See Details.}

\item{pcs}{Vector specifying which principal coordinates to plot. Must be of
length two unless \code{D3 = TRUE}.}

\item{label}{Label data points by sample name? Defaults to \code{FALSE}
unless \code{group} and \code{covar} are both \code{NULL}. If \code{TRUE},
then plot can render at most one phenotypic feature.}

\item{pal_group}{String specifying the color palette to use if \code{group}
is non-\code{NULL}, or a vector of such strings with length equal to the
number of vectors passed to \code{group}. Options include \code{"ggplot"},
all qualitative color schemes available in \code{RColorBrewer}, and the
complete collection of \code{\href{http://bit.ly/2bxnuGB}{ggsci}} palettes.
Alternatively, a character vector of colors with length equal to the
cumulative number of levels in \code{group}.}

\item{pal_covar}{String specifying the color palette to use if \code{covar}
is non-\code{NULL}, or a vector of such strings with length equal to the
number of vectors passed to \code{covar}. Options include all sequential
color schemes available in \code{RColorBrewer}. Alternatively, a
character vector of colors representing a smooth gradient, or a list of
such vectors with length equal to the number of continuous variables to
visualize.}

\item{size}{Point size.}

\item{alpha}{Point transparency.}

\item{title}{Optional plot title.}

\item{legend}{Legend position. Must be one of \code{"bottom"}, \code{"left"},
\code{"top"}, \code{"right"}, \code{"bottomright"}, \code{"bottomleft"},
\code{"topleft"}, or \code{"topright"}.}

\item{hover}{Show sample name by hovering mouse over data point? If \code{
TRUE}, the plot is rendered in HTML and will either open in your browser's
graphic display or appear in the RStudio viewer.}

\item{D3}{Render plot in three dimensions?}
}
\description{
This function plots a low-dimensional projection of an omic data matrix using
multi-dimensional scaling.
}
\details{
MDS is an iterative algorithm for embedding high-dimensional manifolds in
two or three dimensions. Classical MDS is implemented by the \code{
\link[stats]{cmdscale}} function, which finds the optimal two-dimensional
projection of a distance matrix by minimizing the strain of the coordinate
mapping (Torgerson, 1958). Nonmetric MDS (NMDS) is implemented by the \code{
\link[vegan]{monoMDS}} function, which uses isotonic regression to find the
monotonic transformation that minimizes the stress of the embedding
(Kruskal, 1964).

MDS requires a distance matrix as input. Available distance measures include:
\code{"euclidean"}, \code{"maximum"}, \code{"manhattan"}, \code{"canberra"},
\code{"minkowski"}, \code{"cosine"}, \code{"pearson"}, \code{"kendall"},
\code{"spearman"}, \code{"bray"}, \code{"kulczynski"}, \code{"jaccard"},
\code{"gower"}, \code{"altGower"}, \code{"morisita"}, \code{"horn"}, \code{
"mountford"}, \code{"raup"}, \code{"binomial"}, \code{"chao"}, \code{"cao"},
\code{"mahalanobis"}, \code{"MI"}, or \code{"KLD"}. Some distance measures
are unsuitable for certain types of data. See \code{\link{dist_mat}} for more
details on these methods and links to documentation for each. Users may also
directly input a distance matrix calculated using some custom method.

If \code{top} is non-\code{NULL}, then data can either be filtered by
probewise variance (\code{filter_method = "common"}) or using the leading
fold change method of Smyth et al. (\code{filter_method = "pairwise"}). In
the latter case, pairwise distances are calculated using only the \code{top}
most differentially expressed probes between the two samples. This method is
appropriate when different molecular pathways are relevant for distinguishing
different pairs of samples. To run MDS on the complete data, set \code{
top = NULL}. This is functionally equivalent to running PCA on the full
matrix when \code{dist = "euclidean"}. See \code{\link{plot_pca}}.
}
\examples{
mat <- matrix(rnorm(1000 * 5), nrow = 1000, ncol = 5)
plot_mds(mat)

library(DESeq2)
dds <- makeExampleDESeqDataSet()
dds <- rlog(dds)
plot_mds(dds, group = colData(dds)$condition)

}
\references{
Cox, T.F. & Cox, M.A.A. (2001). \emph{Multidimensional Scaling}. Second
edition. Chapman and Hall.

Kruskal, J.B. (1964).
\href{http://bit.ly/2umJmz6}{Multidimensional scaling by optimizing goodness
of fit to a nonmetric hypothesis}. \emph{Psychometrika}, \emph{29}(1): 1-27.

Ritchie, M.E., Phipson, B., Wu, D., Hu, Y., Law, C.W., Shi, W., & Smyth, G.K.
(2015).
\href{https://www.ncbi.nlm.nih.gov/pubmed/25605792}{limma powers differential
expression analyses for RNA-sequencing and microarray studies}. \emph{Nucleic
Acids Res.}, \emph{43}(7): e47.

Torgerson, W.S. (1958). \emph{Theory and Methods of Scaling}. New York:
Wiley.
}
\seealso{
\code{\link[limma]{plotMDS}}, \code{\link{plot_pca}}
}
