% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mds.R
\name{plot_mds}
\alias{plot_mds}
\title{MDS Plot}
\usage{
plot_mds(dat, group = NULL, covar = NULL, top = 500,
  filter_method = "pairwise", pcs = c(1, 2), label = FALSE,
  pal_group = "npg", pal_covar = "blues", title = NULL,
  legend = "right", hover = FALSE, D3 = FALSE)
}
\arguments{
\item{dat}{Omic data matrix or matrix-like object with rows corresponding to
probes and columns to samples. It is strongly recommended that data be
filtered and normalized prior to plotting. Raw counts stored in \code{
\link[edgeR]{DGEList}} or \code{\link[DESeq2]{DESeqDataSet}} objects are
automatically extracted and transformed to the log2-CPM scale, with a
warning.}

\item{group}{Optional character or factor vector of length equal to sample
size, or up to two such vectors organized into a list or data frame. Supply
legend title(s) by passing a named list or data frame.}

\item{covar}{Optional continuous covariate. If non-\code{NULL}, then plot can
render at most one \code{group} variable. Supply legend title by passing
a named list or data frame.}

\item{top}{Optional number (if > 1) or proportion (if < 1) of top probes to
be used for MDS.}

\item{filter_method}{String specifying whether to apply a pairwise or common
filter if \code{top} is non-\code{NULL}. See Details.}

\item{pcs}{Vector specifying which principal coordinates to plot. Must be of
length two unless \code{D3 = TRUE}.}

\item{label}{Label data points by sample name? Defaults to \code{FALSE}
unless \code{group} and \code{covar} are both \code{NULL}. If \code{TRUE},
then plot can render at most one phenotypic feature.}

\item{pal_group}{String specifying the color palette to use if \code{group}
is non-\code{NULL}. Options include \code{"ggplot"}, as well as the
complete collection of \code{
\href{https://cran.r-project.org/web/packages/ggsci/vignettes/ggsci.html}{
ggsci}} palettes, which can be identified by name (e.g., \code{"npg"},
\code{"aaas"}, etc.). Alternatively, any character vector of colors with
length equal to the number of levels in \code{group}.}

\item{pal_covar}{String specifying the color palette to use if \code{covar}
is non-\code{NULL}. Options include \code{"blues"}, \code{"greens"}, \code{
"purples"}, \code{"greys"}, \code{"oranges"}, and \code{"reds"}.
Alternatively, any character vector of colors representing a smooth
gradient.}

\item{title}{Optional plot title.}

\item{legend}{Legend position. Must be one of \code{"right"}, \code{
"left"}, \code{"top"}, \code{"bottom"}, \code{"topright"}, \code{
"topleft"}, \code{"bottomright"}, or \code{"bottomleft"}.}

\item{hover}{Show sample name by hovering mouse over data point? If \code{
TRUE}, the plot is rendered in HTML and will either open in your browser's
graphic display or appear in the RStudio viewer.}

\item{D3}{Render plot in three dimensions?}
}
\description{
This function plots a low-dimensional projection of an omic data matrix using
multi-dimensional scaling.
}
\details{
This function plots the samples of an omic data matrix in a two- or
three-dimensional principal coordinate subspace. MDS is an easy and popular
method for unsupervised cluster detection. It can also aid in spotting
potential outliers, and generally helps to visualize the latent structure of
a data set.

If \code{top} is non-\code{NULL}, then data can either be filtered by
probewise variance (\code{filter_method = "common"}) or using the leading
fold change method of Smyth et al. (\code{filter_method = "pairwise"}). In
the latter case, pairwise Euclidean distances are calculated using only the
\code{top} most differentially expressed probes between the two samples. This
method is appropriate when different molecular pathways are relevant for
distinguishing different pairs of samples.

To run MDS on the complete data, set \code{top = NULL}. This is functionally
equivalent to running PCA on the full matrix. See \code{\link{plot_pca}}.
}
\examples{
mat <- matrix(rnorm(1000 * 5), nrow = 1000, ncol = 5)
plot_mds(mat)

library(DESeq2)
dds <- makeExampleDESeqDataSet()
dds <- rlog(dds)
plot_mds(dds, group = colData(dds)$condition)

}
\references{
Cox, T.F. & Cox, M.A.A. (2001). \emph{Multidimensional Scaling}. Second
edition. Chapman and Hall.

Ritchie, M.E., Phipson, B., Wu, D., Hu, Y., Law, C.W., Shi, W., & Smyth, G.K.
(2015).
\href{https://www.ncbi.nlm.nih.gov/pubmed/25605792}{limma powers differential
expression analyses for RNA-sequencing and microarray studies}. \emph{Nucleic
Acids Res.}, emph{43}(7): e47.

Torgerson, W.S. (1958). \emph{Theory and Methods of Scaling}. New York:
Wiley.
}
\seealso{
\code{\link[limma]{plotMDS}}, \code{\link{plot_pca}}, \code{\link{plot_tsne}}
}
