% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mean_var.R
\name{plot_mean_var}
\alias{plot_mean_var}
\title{Plot the mean-variance trend of a gene expression matrix}
\usage{
plot_mean_var(dat, type, ptsize = 0.25, main = NULL, hover = FALSE,
  knitr = FALSE)
}
\arguments{
\item{dat}{Matrix of log expression intensities or raw counts.}

\item{type}{String specifying data type. Must be one of either
\code{"microarray"} or \code{"RNA-seq"}.}

\item{main}{Optional plot title.}

\item{hover}{Show probe name by hovering mouse over data point? If \code{TRUE},
the plot is rendered in HTML and will either open in your browser's graphic
display or appear in the RStudio viewer. The plot can also be embedded in an
HTML doc using Rmarkdown so long as \code{knitr = TRUE} and the code chunk
option \code{plotly} is also set to \code{TRUE}.}

\item{knitr}{Set this to \code{TRUE} if you want to embed a plotly object (viz.,
the \code{plot_mean_var} output when \code{hover = TRUE}) in an HTML doc. Make
sure to set \code{plotly = TRUE} in the corresponding code chunk options.}
}
\description{
Plot the mean-variance trend of a gene expression matrix
}
\details{
This function plots each gene's mean expression against either the
logarithm of its standard deviation (if \code{type = "microarray"}) or
its square root (if \code{type = "RNA-seq"}). A lowess curve is additionally
fit to the data.
}
\examples{
mat <- matrix(rnorm(5000), nrow = 1000, ncol = 5)
plot_mean_var(mat, type = "microarray")

mat <- matrix(rnbinom(5000, mu = 4, size = 1), nrow = 1000, ncol = 5)
mat <- cpm(y, log = TRUE)
plot_mean_var(mat, type = "RNA-seq")

}

