% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mean_var.R
\name{plot_mean_var}
\alias{plot_mean_var}
\title{Mean-Variance Plot}
\usage{
plot_mean_var(dat, trans, ptsize = 0.25, main = NULL, hover = FALSE,
  probes = NULL)
}
\arguments{
\item{dat}{Omic data matrix with rows corresponding to probes and columns
to samples. Data are presumed to be normalized prior to visualization.}

\item{trans}{Data transformation to be applied to the probewise standard
deviations. Must be one of either \code{"log"} or \code{"sqrt"}. The former
is recommended for microarrays or any other platform in which data are
approximately log-normally distributed. The latter is recommended for
sequencing and count data.}

\item{main}{Optional plot title.}

\item{hover}{Show probe name by hovering mouse over data point? If \code{TRUE},
the plot is rendered in HTML and will either open in your browser's graphic
display or appear in the RStudio viewer.}

\item{probes}{String specifying the name of the column in which to find the probe
identifiers, assuming they aren't \code{rownames(dat)}. Only relevant if
\code{hover = TRUE}.}
}
\description{
This function plots probewise means vs. standard deviations, following
the appropriate data transformation.
}
\details{
Mean-variance plots are a quick and easy way to visualize distributions
of probewise expression or methylation values across samples. A lowess curve
is additionally fit to the data.
}
\examples{
mat <- matrix(rnorm(5000), nrow = 1000, ncol = 5)
plot_mean_var(mat, trans = "log")

library(edgeR)
mat <- matrix(rnbinom(5000, mu = 4, size = 1), nrow = 1000, ncol = 5)
mat <- cpm(y, log = TRUE)
plot_mean_var(mat, trans = "sqrt")

}
