% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mean_var.R
\name{plot_mean_var}
\alias{plot_mean_var}
\title{Plot the mean-variance trend of a gene expression matrix}
\usage{
plot_mean_var(dat, type, main = NULL)
}
\arguments{
\item{dat}{Matrix of log expression intensities or raw counts.}

\item{type}{String specifying data type. Must be one of either 
\code{"microarray"} or \code{"RNA-seq"}.}

\item{main}{Optional plot title.}
}
\description{
Plot the mean-variance trend of a gene expression matrix
}
\details{
This function plots each gene's mean expression against either the 
logarithm of its standard deviation (if \code{type = "microarray"}) or
its square root (if \code{type = "RNA-seq"}). A lowess curve is additionally 
fit to the data.
}
\examples{
mat <- matrix(rnorm(5000), nrow = 1000, ncol = 5)
plot_mean_var(mat, type = "microarray")

mat <- matrix(rnbinom(5000, mu = 4, size = 1), nrow = 1000, ncol = 5)
plot_mean_var(mat, type = "RNA-seq")

}

