% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mv.R
\name{plot_mv}
\alias{plot_mv}
\alias{plot_mv.MArrayLM}
\alias{plot_mv.DGEList}
\alias{plot_mv.DGELM}
\alias{plot_mv.DESeqDataSet}
\alias{plot_mv.DESeqTransform}
\alias{plot_mv.default}
\title{Mean-Variance Plot}
\usage{
plot_mv(
  dat,
  trans = "rank",
  title = "Mean-Variance Plot",
  legend = "right",
  ...
)

\method{plot_mv}{MArrayLM}(
  dat,
  trans = "log",
  span = 0.5,
  title = "Mean-Variance Plot",
  xlab = NULL,
  ylab = NULL,
  legend = "right",
  hover = FALSE
)

\method{plot_mv}{DGEList}(
  dat,
  design = NULL,
  trans = "sqrt",
  span = 0.5,
  title = "Mean-Variance Plot",
  xlab = NULL,
  ylab = NULL,
  legend = "right",
  hover = FALSE
)

\method{plot_mv}{DGELM}(
  dat,
  trans = "sqrt",
  span = 0.5,
  title = "Mean-Variance Plot",
  xlab = NULL,
  ylab = NULL,
  legend = "right",
  hover = FALSE
)

\method{plot_mv}{DESeqDataSet}(
  dat,
  resid = FALSE,
  trans = "rank",
  span = 0.5,
  title = "Mean-Variance Plot",
  xlab = NULL,
  ylab = NULL,
  legend = "right",
  hover = FALSE
)

\method{plot_mv}{DESeqTransform}(
  dat,
  trans = "rank",
  span = 0.5,
  title = "Mean-Variance Plot",
  xlab = NULL,
  ylab = NULL,
  legend = "right",
  hover = FALSE
)

\method{plot_mv}{default}(
  dat,
  trans = "rank",
  span = 0.5,
  title = "Mean-Variance Plot",
  xlab = NULL,
  ylab = NULL,
  legend = "right",
  hover = FALSE
)
}
\arguments{
\item{dat}{Omic data matrix or matrix-like object with rows corresponding to
probes and columns to samples. Data are presumed to be filtered and
normalized prior to visualization. Alternatively, any object from the
\code{limma}, \code{edgeR}, or \code{DESeq2} pipelines. See Details.}

\item{trans}{Data transformation to be applied to probewise means (if \code{
trans = "rank"}) or standard deviations (if \code{trans = "log"} or \code{
"sqrt"}). Not all transformations are appropriate for all data types. See
Details.}

\item{title}{Optional plot title.}

\item{legend}{Legend position. Must be one of \code{"bottom"}, \code{"left"},
\code{"top"}, \code{"right"}, \code{"bottomright"}, \code{"bottomleft"},
\code{"topleft"}, or \code{"topright"}.}

\item{span}{Width of the LOWESS smoothing window as a proportion.}

\item{xlab}{Optional label for x-axis.}

\item{ylab}{Optional label for y-axis.}

\item{hover}{Show probe name by hovering mouse over data point? If \code{
TRUE}, the plot is rendered in HTML and will either open in your browser's
graphic display or appear in the RStudio viewer. Probe names are extracted
from \code{dat}.}

\item{design}{Optional design matrix with rows corresponding to samples and
columns to coefficients to be estimated. Only relevant for \code{
\link[edgeR]{DGEList}} objects. See Details.}

\item{resid}{Should y-axis plot genewise variance or residual variance? Only
relevant for \code{DESeqDataSet}s that have been fit with a negative
binomial model. See Details.}
}
\description{
This function visualizes the mean-variance relationship of omic data before
or after modeling.
}
\details{
Mean-variance (MV) plots are a quick and easy way to visualize the
relationship between the first two moments of probewise data distributions.
When used prior to modeling, they may help better understand the internal
structure of the data and inspect for potential outliers. The effects of
filtering and transformations can also be readily evaluated. When applied
after modeling, MV plots help assess the assumptions of the regression.

If \code{dat} is a matrix or object inheriting from the \code{
\link[Biobase]{ExpressionSet}} class, then it is presumed to be filtered and
normalized prior to plotting. For count data, this means undergoing some sort
of variance stabilizing transformation, such as \code{\link[edgeR]{cpm}}
(with \code{log = TRUE}), \code{\link[DESeq2]{vst}}, \code{\link[DESeq2]{
rlog}}, etc. If \code{dat} is of class \code{\link[limma]{MArrayLM}} or
\code{\link[edgeR]{DGEGLM}}, then the y-axis represents residual variance.
For \code{\link[DESeq2]{DESeqDataSet}}s, the choice of whether to plot
variance or residual variance is determined by the \code{resid} argument. For
data that has been fit with a negative binomial model, a residual matrix is
generated by subtracting the normalized signal matrix from the normalized
counts on the log2-CPM scale.

Default setting for the \code{trans} argument vary by S3 class. When \code{
trans = 'rank'}, rank-transformed means are plotted against standard
deviations for each probe. This is a reasonable choice for most data types.
It is also common in the microarray literature to plot probewise means
against log-transformed standard deviations, while the \code{limma} authors
recommend the square-root transform for count data on the log2-CPM scale.

Note that if passing a matrix of log2-CPM transformed counts, the plot will
not be exactly the same as the output of \code{\link[limma]{voom}} when
\code{plot = TRUE}. This is because the \code{voom} transform involves an
initial modeling step in which probewise means are plotted against
square-root transformed residual standard deviations on the log2-CPM scale.
See \code{\link{plot_voom}} to recreate the voom transformation plot in
\code{bioplotr} style.

\code{plot_mv} fits a LOWESS curve to the points with a smoothing window
given by \code{span}. If \code{dat} is an \code{MArrayLM} object with
standard errors moderated by \code{eBayes}, then prior variance will also be
plotted as either a horizontal line or a smooth curve, depending on whether a
global or intensity-dependent prior was used. If robust empirical Bayes was
used to create \code{dat}, outlier variances are highlighted. See \code{
\link[limma]{squeezeVar}} for more details.
}
\examples{
mat <- matrix(rnorm(1000 * 10), nrow = 1000, ncol = 10)
plot_mv(mat, trans = "log")

library(limma)
grp <- rep(c("ctl", "trt"), each = 5)
des <- model.matrix(~ grp)
fit <- eBayes(lmFit(mat, des), trend = TRUE)
plot_mv(fit)

}
\references{
Huber, W., Hedebreck, A., Sültmann, H., Poustka, A. & Vingron, M. (2002).
\href{https://www.ncbi.nlm.nih.gov/pubmed/12169536}{Variance Stabilization
Applied to Microarray Data Calibration and to the Quantification of
Differential Expression}. \emph{Bioinformatics}, \emph{18}(1), S96-2104.

Sartor, M.A., Tomlinson, C.R., Wesselkamper, S.C., Sivaganesan, S., Leikauf,
G.D. & Medvedovic, M. (2006).
\href{http://bmcbioinformatics.biomedcentral.com/articles/10.1186/1471-2105-7-538}{
Intensity-based hierarchical Bayes method improves testing for differentially
expressed genes in microarray experiments}. \emph{BMC Bioinformatics},
\strong{7}: 538.

Law, C.W., Chen, Y., Shi, W., & Smyth, G.K. (2014).
\href{https://genomebiology.biomedcentral.com/articles/10.1186/gb-2014-15-2-r29}{
voom: precision weights unlock linear model analysis tools for RNA-seq read
counts}. \emph{Genome Biology}, \strong{15}: R29.
}
\seealso{
\code{\link[edgeR]{plotMeanVar}}, \code{\link[edgeR]{dglmStdResid}},
\code{\link[vsn]{meanSdPlot}}, \code{\link[limma]{plotSA}},
\code{\link[limma]{voom}}
}
