% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca.R
\name{plot_pca}
\alias{plot_pca}
\title{Plot the principal components of an omic data matrix}
\usage{
plot_pca(dat, group = NULL, label = FALSE, main = NULL,
  legend = "outside", hover = FALSE, D3 = FALSE, knitr = FALSE)
}
\arguments{
\item{dat}{Omic data matrix with rows corresponding to probes and columns
to samples. It is strongly recommended that data be filtered and
normalized prior to running PCA.}

\item{group}{Optional factor or character vector of length equal to sample size.
Levels are used to color and shape points. If supplied, legend title defaults to
"Group". Override this feature by passing a named list instead.}

\item{label}{Label data points by sample name? Defaults to \code{FALSE} unless
\code{group = NULL}.}

\item{main}{Optional plot title.}

\item{legend}{Legend position. Must be one of \code{"outside",
"bottomleft", "bottomright", "topleft",} or \code{"topright"}.}

\item{hover}{Show sample name by hovering mouse over data point? If \code{TRUE},
the plot is rendered in HTML and will either open in your browser's graphic
display or appear in the RStudio viewer.}

\item{D3}{Render the plot in three dimensions?}
}
\description{
Plot the principal components of an omic data matrix
}
\details{
This function plots the samples of an omic data matrix in a two- or three-
dimensional principal component subspace. The numbers printed along the axis
labels indicate the percentage of variance explained by each component. PCA is an
easy and popular method for unsupervised cluster detection. It can also aid in
spotting potential outliers and generally helps to visualize the latent
structure of a data set.
}
\examples{
mat <- matrix(rnorm(5000), nrow = 1000, ncol = 5)
plot_pca(mat)

library(DESeq2)
mat <- cbind(matrix(rnbinom(5000, mu = 4, size = 1), nrow = 1000, ncol = 5),
             matrix(rnbinom(5000, mu = 4, size = 10), nrow = 1000, ncol = 5))
mat <- rlog(mat)
grp <- rep(c("A", "B"), each = 5)
plot_pca(mat, group = grp)

}

