% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca.R
\name{plot_pca}
\alias{plot_pca}
\title{PCA Plot}
\usage{
plot_pca(dat, group = NULL, covar = NULL, top = NULL, pcs = c(1L,
  2L), label = FALSE, pal_group = "npg", pal_covar = "Blues",
  title = NULL, legend = "right", hover = FALSE, D3 = FALSE)
}
\arguments{
\item{dat}{Omic data matrix or matrix-like object with rows corresponding to
probes and columns to samples. It is strongly recommended that data be
filtered and normalized prior to plotting. Raw counts stored in \code{
\link[edgeR]{DGEList}} or \code{\link[DESeq2]{DESeqDataSet}} objects are
automatically extracted and transformed to the log2-CPM scale, with a
warning.}

\item{group}{Optional character or factor vector of length equal to sample
size, or up to two such vectors organized into a list or data frame. Supply
legend title(s) by passing a named list or data frame.}

\item{covar}{Optional continuous covariate. If non-\code{NULL}, then plot can
render at most one \code{group} variable. Supply legend title by passing
a named list or data frame.}

\item{top}{Optional number (if > 1) or proportion (if < 1) of most variable
probes to be used for PCA.}

\item{pcs}{Vector specifying which principal components to plot. Must be of
length two unless \code{D3 = TRUE}.}

\item{label}{Label data points by sample name? Defaults to \code{FALSE}
unless \code{group} and \code{covar} are both \code{NULL}. If \code{TRUE},
then plot can render at most one phenotypic feature.}

\item{pal_group}{String specifying the color palette to use if \code{group}
is non-\code{NULL}, or a vector of such strings with length equal to the
number of vectors passed to \code{group}. Options include \code{"ggplot"},
all qualitative color schemes available in \code{RColorBrewer}, and the
complete collection of \code{\href{http://bit.ly/2bxnuGB}{ggsci}} palettes.
Alternatively, a character vector of colors with length equal to the
cumulative number of levels in \code{group}.}

\item{pal_covar}{String specifying the color palette to use if \code{covar}
is non-\code{NULL}, or a vector of such strings with length equal to the
number of vectors passed to \code{covar}. Options include all sequential
color schemes available in \code{RColorBrewer}. Alternatively, a
character vector of colors representing a smooth gradient, or a list of
such vectors with length equal to the number of continuous variables to
visualize.}

\item{title}{Optional plot title.}

\item{legend}{Legend position. Must be one of \code{"bottom"}, \code{"left"},
\code{"top"}, \code{"right"}, \code{"bottomright"}, \code{"bottomleft"},
\code{"topleft"}, or \code{"topright"}.}

\item{hover}{Show sample name by hovering mouse over data point? If \code{
TRUE}, the plot is rendered in HTML and will either open in your browser's
graphic display or appear in the RStudio viewer.}

\item{D3}{Render plot in three dimensions?}
}
\description{
This function plots a low-dimensional projection of an omic data matrix using
principal component analysis.
}
\details{
This function plots the samples of an omic data matrix in a two- or
three-dimensional principal component subspace. Axis labels include the
percentage of variance explained by each component. PCA is an easy and
popular projection method that can aid in identifying clusters, detecting
outliers, and visualizing the latent structure of a dataset.

By default, \code{plot_pca} performs singular value decomposition on the
complete \code{dat} matrix. Limit the PCA to only the most variable probes by
using the \code{top} argument.
}
\examples{
mat <- matrix(rnorm(1000 * 5), nrow = 1000, ncol = 5)
plot_pca(mat)

library(DESeq2)
dds <- makeExampleDESeqDataSet()
dds <- rlog(dds)
plot_pca(dds, group = colData(dds)$condition)

}
\references{
Hotelling, H. (1933).
\href{http://psycnet.apa.org/journals/edu/24/6/417/}{Analysis of a complex of
variables into principal components}. \emph{Journal of Educational
Psychology}, \emph{24}(6): 414:441.

Pearson, K. (1901).
\href{http://www.tandfonline.com/doi/abs/10.1080/14786440109462720}{On lines
and planes of closest fit to systems of points in space}. \emph{Philosophical
Magazine}, \emph{2}(11): 559–572.
}
\seealso{
\code{\link[DESeq2]{plotPCA}}, \code{\link{plot_mds}},
\code{\link{plot_tsne}}
}
