% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca.R
\name{plot_pca}
\alias{plot_pca}
\title{PCA Plot}
\usage{
plot_pca(dat, group = NULL, covar = NULL, top = NULL, pcs = c(1, 2),
  label = FALSE, title = NULL, legend = "outside", hover = FALSE,
  D3 = FALSE)
}
\arguments{
\item{dat}{Omic data matrix or matrix-like object with rows corresponding to
probes and columns to samples. It is strongly recommended that data be
normalized and filtered prior to running PCA. For count data, this means
undergoing some sort of variance stabilizing transformation, such as
\code{\link{lcpm}, \link[DESeq2]{vst}, \link[DESeq2]{rlog}}, etc.}

\item{group}{Optional character or factor vector of length equal to sample size,
or up to two such vectors organized into a list or data frame. Supply legend
title(s) by passing a named list or data frame.}

\item{covar}{Optional continuous covariate. If non-\code{NULL}, then function can
take at most one \code{group} variable. Supply legend title by passing a named
list or data frame.}

\item{top}{Optional number (if > 1) or proportion (if < 1) of most variable probes
to be used for PCA.}

\item{pcs}{Vector specifying which principal components to plot. Must be of length
two unless \code{D3 = TRUE}.}

\item{label}{Label data points by sample name? Defaults to \code{FALSE} unless
\code{covar = NULL}. If \code{TRUE}, then plot can render at most one covariate.}

\item{title}{Optional plot title.}

\item{legend}{Legend position. Must be one of \code{"outside", "bottomleft",
"bottomright", "topleft",} or \code{"topright"}.}

\item{hover}{Show sample name by hovering mouse over data point? If \code{TRUE},
the plot is rendered in HTML and will either open in your browser's graphic
display or appear in the RStudio viewer.}

\item{D3}{Render plot in three dimensions?}
}
\description{
This function plots a low-dimensional projection of an omic data matrix using
principal component analysis.
}
\details{
This function plots the samples of an omic data matrix in a two- or
three-dimensional principal component subspace. Axis labels include the percentage
of variance explained by each component. PCA is an easy and popular method for
unsupervised cluster detection. It can also aid in spotting potential outliers,
and generally helps to visualize the latent structure of a data set.

By default, \code{plot_pca} performs singular value decomposition on the complete
\code{dat} matrix. Limit the PCA to only the most variable probes by using the
\code{top} argument.
}
\examples{
mat <- matrix(rnorm(1000 * 5), nrow = 1000, ncol = 5)
plot_pca(mat)

library(DESeq2)
dds <- makeExampleDESeqDataSet()
dds <- rlog(dds)
plot_pca(mat, group = colData(dds)$condition)

}
\references{
Hotelling, H. (1933). \href{http://psycnet.apa.org/journals/edu/24/6/417/}{Analysis
of a complex of variables into principal components}. \emph{Journal of Educational
Psychology}, \emph{24}(6): 414:441.

Pearson, K. (1901). \href{http://www.tandfonline.com/doi/abs/10.1080/14786440109462720}{On
lines and planes of closest fit to systems of points in space}. \emph{Philosophical
Magazine}, \emph{2}(11): 559–572.
}
\seealso{
\code{\link[DESeq2]{plotPCA}, \link{plot_mds}, \link{plot_tsne}}
}
