% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_qq.R
\name{plot_qq}
\alias{plot_qq}
\title{Q-Q Plot}
\usage{
plot_qq(dat, ptsize = 0.25, main = NULL, legend = "outside",
  hover = FALSE, probes = NULL)
}
\arguments{
\item{dat}{Data frame or matrix representing the results of a test for
differential expression or methylation, such as the output of a call to
\code{limma::topTable, edgeR::topTags}, or \code{DESeq2::results}.
Alternatively, any object with a column for \emph{p}-values.}

\item{ptsize}{Size of data points in the plot.}

\item{main}{Optional plot title.}

\item{hover}{Show probe name by hovering mouse over data point? If \code{TRUE},
the plot is rendered in HTML and will either open in your browser's graphic
display or appear in the RStudio viewer.}

\item{probes}{String specifying the name of the column in which to find the probe
identifiers, assuming they are not \code{rownames(dat)}. Only relevant if
\code{hover = TRUE}.}
}
\description{
This function plots expected vs. observed quantiles of a \emph{p}-value
distribution following -log10 transform.
}
\details{
Q-Q plots are a common way to visually assess the applicability of a statistical
test to a given data set. If the black points deviate too sharply from the red
line, especially at low expected values of -log10(\emph{p}), then it suggests a
violation of the assumptions upon which the test was based.
}
\examples{
df <- data.frame(p.value = runif(10000))
plot_qq(df)

library(limma)
DE_genes <- cbind(matrix(rnorm(50 * 5, mean = 5), nrow = 50, ncol = 5),
                  matrix(rnorm(50 * 5), nrow = 50, ncol = 5))
eset <- rbind(DE_genes, matrix(rnorm(4950 * 10), nrow = 4950, ncol = 10))
treat <- rep(c("A", "B"), each = 5)
des <- model.matrix(~ treat)
fit <- eBayes(lmFit(eset, des))
top <- topTable(fit, number = Inf)
plot_qq(top)

}
