% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_qq.R
\name{plot_qq}
\alias{plot_qq}
\title{Create QQ plot of expected vs. observed -log10 \emph{p}-values}
\usage{
plot_qq(dat, ptsize = 0.25, main = NULL, legend = "outside",
  hover = FALSE, probes = NULL)
}
\arguments{
\item{dat}{Data frame or matrix representing the results of a test for
differential expression or methylation, such as the output of a call to
\code{limma::topTable, edgeR::topTags}, or \code{DESeq2::results}.
Alternatively, any object with a column for \emph{p}-values.}

\item{ptsize}{Size of data points in the plot.}

\item{main}{Optional plot title.}

\item{legend}{Legend position. Must be one of \code{"outside",
"bottomleft", "bottomright", "topleft",} or \code{"topright"}.}

\item{hover}{Show probe name by hovering mouse over data point? If \code{TRUE},
the plot is rendered in HTML and will either open in your browser's graphic
display or appear in the RStudio viewer.}

\item{probes}{String specifying the name of the column in which to find the probe
identifiers. Only relevant if \code{hover = TRUE}.}
}
\description{
Create QQ plot of expected vs. observed -log10 \emph{p}-values
}
\details{
This function displays a scatterplot of the expected vs. observed quantiles
of a \emph{p}-value distribution following -log10 transform. If the black points
deviate too sharply from the red line, especially at low expected values of
-log10(\emph{p}), then it suggests a violation of the statistical assumptions
upon which the test was based.
}
\examples{
df <- data.frame(p.value = runif(10000))
plot_qq(df)

library(limma)
DE_genes <- cbind(matrix(rnorm(250, 5, 1), nrow = 50, ncol = 5),
                  matrix(rnorm(250), nrow = 50, ncol = 5))
mat <- rbind(DE_genes, matrix(rnorm(45500), nrow = 4550, ncol = 10))
treat <- gl(n = 2, k = 5, labels = c("A", "B"))
des <- model.matrix(~ treat)
fit <- eBayes(lmFit(mat, des))
top <- topTable(fit, number = Inf)
plot_qq(top)

}

