% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_roc.R
\name{plot_roc}
\alias{plot_roc}
\title{Plot ROC curves for one or several classifiers}
\usage{
plot_roc(obs, pred, main = NULL, legend = "bottomright", hover = FALSE)
}
\arguments{
\item{obs}{Vector of observed outcomes. Must be dichotomous. Can be numeric,
logical, character, or factor. If numeric, \code{obs} must be coded \code{1}
or \code{0}. If character or factor, a warning will be issued clarifying that
the first level is assumed to be the reference.}

\item{pred}{Vector of predicted values or a list of such vectors, optionally named.
Must be numeric. Common examples include the probabilities output by a logistic
model, or the expression levels of a particular biomarker.}

\item{main}{Optional plot title.}

\item{legend}{Legend position. Must be one of \code{"outside", "bottomleft",
"bottomright", "topleft",} or \code{"topright"}.}

\item{hover}{Show predictor name by hovering mouse over ROC curve? If \code{TRUE},
the plot is rendered in HTML and will either open in your browser's graphic
display or appear in the RStudio viewer.}
}
\description{
Plot ROC curves for one or several classifiers
}
\details{
This function plots one or several receiver operating characteristic (ROC) curves.
ROC curves plot the false positive rate (1 - specificity) vs. the true positive
rate (sensitivity) for a given classifier and vector of observations. The area
under the ROC curve (AUC) is a common performance metric for binary classifiers.
}
\examples{
y <- rbinom(300, size = 1, prob = 0.5)
x <- rnorm(300, mean = y, sd = 0.5)
plot_roc(obs = y, pred = x)

x2 <- rnorm(300, mean = y, sd = 2)
plot_roc(obs = y, pred = list("x1" = x, "x2" = x2))

}

