% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sammon.R
\name{plot_sammon}
\alias{plot_sammon}
\title{Sammon Map}
\usage{
plot_sammon(dat, group = NULL, covar = NULL, dist = "euclidean", p = 2,
  top = 500, filter_method = "pairwise", dims = c(1, 2), label = FALSE,
  pal_group = "npg", pal_covar = "blues", title = NULL,
  legend = "right", hover = FALSE, D3 = FALSE)
}
\arguments{
\item{dat}{Omic data matrix or matrix-like object with rows corresponding to
probes and columns to samples. It is strongly recommended that data be
filtered and normalized prior to plotting. Raw counts stored in \code{
\link[edgeR]{DGEList}} or \code{\link[DESeq2]{DESeqDataSet}} objects are
automatically extracted and transformed to the log2-CPM scale, with a
warning.}

\item{group}{Optional character or factor vector of length equal to sample
size, or up to two such vectors organized into a list or data frame. Supply
legend title(s) by passing a named list or data frame.}

\item{covar}{Optional continuous covariate. If non-\code{NULL}, then plot can
render at most one \code{group} variable. Supply legend title by passing
a named list or data frame.}

\item{dist}{Distance measure to be used. Supports all methods available in
\code{\link[stats]{dist}} and \code{\link[vegan]{vegdist}}, as well as
those implemented in the \code{bioDist} package. See Details.}

\item{p}{Power of the Minkowski distance.}

\item{top}{Optional number (if > 1) or proportion (if < 1) of top probes to
be used for mapping.}

\item{filter_method}{String specifying whether to apply a \code{"pairwise"}
or \code{"common"} filter if \code{top} is non-\code{NULL}. See Details.}

\item{dims}{Vector specifying which dimensions to plot. Must be of length
two unless \code{D3 = TRUE}.}

\item{label}{Label data points by sample name? Defaults to \code{FALSE}
unless \code{group} and \code{covar} are both \code{NULL}. If \code{TRUE},
then plot can render at most one phenotypic feature.}

\item{pal_group}{String specifying the color palette to use if \code{group}
is non-\code{NULL}. Options include \code{"ggplot"}, as well as the
complete collection of \code{
\href{https://cran.r-project.org/web/packages/ggsci/vignettes/ggsci.html}{
ggsci}} palettes, which can be identified by name (e.g., \code{"npg"},
\code{"aaas"}, etc.). Alternatively, any character vector of colors with
length equal to the number of levels in \code{group}.}

\item{pal_covar}{String specifying the color palette to use if \code{covar}
is non-\code{NULL}. Options include \code{"blues"}, \code{"greens"}, \code{
"purples"}, \code{"greys"}, \code{"oranges"}, and \code{"reds"}.
Alternatively, any character vector of colors representing a smooth
gradient.}

\item{title}{Optional plot title.}

\item{legend}{Legend position. Must be one of \code{"right"}, \code{
"left"}, \code{"top"}, \code{"bottom"}, \code{"topright"}, \code{
"topleft"}, \code{"bottomright"}, or \code{"bottomleft"}.}

\item{hover}{Show sample name by hovering mouse over data point? If \code{
TRUE}, the plot is rendered in HTML and will either open in your browser's
graphic display or appear in the RStudio viewer.}

\item{D3}{Render plot in three dimensions?}
}
\description{
This function plots a low-dimensional projection of an omic data matrix using
Sammon mapping.
}
\details{
Sammon mapping is a variant of nonmetric MDS with a cost function designed
to better preserve local structure.

The projection is calculated using the \code{MASS::\link[MASS]{sammon}}
function, which takes a distance matrix as input. Available distance
measures include: \code{"euclidean"}, \code{"maximum"}, \code{"manhattan"},
\code{"canberra"}, \code{"minkowski"}, \code{"cosine"}, \code{"pearson"},
\code{"kendall"}, \code{"spearman"}, \code{"bray"}, \code{"kulczynski"},
\code{"jaccard"}, \code{"gower"}, \code{"altGower"}, \code{"morisita"},
\code{"horn"}, \code{"mountford"}, \code{"raup"}, \code{"binomial"}, \code{
"chao"}, \code{"cao"}, \code{"mahalanobis"}, \code{"MI"}, or \code{"KLD"}.
Some distance measures are unsuitable for certain types of data. See \code{
\link{dist_mat}} for more details on these methods and links to documentation
on each.

The \code{top} argument optionally filters data using either probewise
variance (if \code{filter_method = "common"}) or the leading fold change
method of Smyth et al. (if \code{filter_method = "pairwise"}). See \code{
\link{plot_mds}} for more details.
}
\examples{
mat <- matrix(rnorm(1000 * 5), nrow = 1000, ncol = 5)
plot_sammon(mat)

library(DESeq2)
dds <- makeExampleDESeqDataSet()
dds <- rlog(dds)
plot_sammon(dds, group = colData(dds)$condition)

}
\references{
Sammon, J.W. (1969).
\href{http://theoval.cmp.uea.ac.uk/~gcc/matlab/sammon/sammon.pdf}{A
non-linear mapping for data structure analysis}. \emph{IEE Trans. Comput.},
\strong{C-18}: 401-409.
}
\seealso{
\code{\link[limma]{plot_mds}}, \code{\link{plot_tsne}}
}
