% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_similarity.R
\name{plot_similarity}
\alias{plot_similarity}
\title{Similarity Matrix Heatmap}
\usage{
plot_similarity(dat, feat = NULL, dist = "euclidean",
  hclustfun = "average", title = NULL)
}
\arguments{
\item{dat}{Omic data matrix or matrix-like object with rows corresponding to
probes and columns to samples. It is strongly recommended that data be
normalized and filtered prior to plotting the sample similarity matrix.
For count data, this means undergoing some sort of variance stabilizing
transformation, such as \code{\link[edgeR]{cpm} (with \code{log = TRUE}),
\link[DESeq2]{vst}, \link[DESeq2]{rlog}}, etc.}

\item{feat}{Optional character, factor, numeric, or logical vector of length
equal to sample size. Alternatively, a data frame or list of such vectors,
optionally named. Values are used to color one or several annotation tracks
atop the heatmap.}

\item{dist}{Distance measure to be used. Currently supports \code{"euclidean",
"pearson", "MI",} or \code{"KLD"}. See Details.}

\item{hclustfun}{The agglomeration method to be used for hierarchical clustering.
See \code{\link[stats]{hclust}} for available options.}

\item{title}{Optional plot title.}
}
\description{
This function displays pairwise distances between samples as a heatmap.
}
\details{
Similarity matrices are a valuable tool for exploratory data analysis. A
hierarchical clustering dendrogram atop the figure helps identify potential
clusters and/or outliers in the data. Annotation tracks can help investigate
associations with phenotypic features.

Different distance measures and agglomeration methods can lead to different
results. The default settings, which perform average linkage hierarchical
clustering on a Euclidean distance matrix, are mathematically straightforward
and commonly used for omic EDA. Complete linkage is also fairly common for
hierarchical clustering, while other options (single linkage, Ward, etc.) are less
informative.

Pearson distance, defined as 1 - the Pearson correlation, is another popular
method for evaluating sample similarity. Mutual information and Kullback-Leibler
divergence are more complicated distance metrics that require some simplifying
assumptions to be efficiently applied to continuous data distributions. See
\code{\link[bioDist]{MIdist}} and \code{\link[bioDist]{KLdist.matrix}}
for more info.
}
\examples{
mat <- matrix(rnorm(5000), nrow = 1000, ncol = 5)
plot_similarity(mat, title = "Nothin' Doin'")

library(DESeq2)
mat <- cbind(matrix(rnbinom(5000, mu = 4, size = 1), nrow = 1000, ncol = 5),
             matrix(rnbinom(5000, mu = 4, size = 10), nrow = 1000, ncol = 5))
mat <- rlog(mat)
grp <- rep(c("A", "B"), each = 5)
plot_similarity(mat, feat = grp, title = "Somethin' Cookin'")

}
