% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_similarity.R
\name{plot_similarity}
\alias{plot_similarity}
\title{Similarity Matrix Heatmap}
\usage{
plot_similarity(
  dat,
  group = NULL,
  covar = NULL,
  dist = "euclidean",
  p = 2L,
  top = NULL,
  filter_method = "pairwise",
  center = FALSE,
  hclustfun = "average",
  pal_group = "npg",
  pal_covar = "Blues",
  pal_tiles = "RdBu",
  title = "Sample Similarity Matrix"
)
}
\arguments{
\item{dat}{Omic data matrix or matrix-like object with rows corresponding to
probes and columns to samples. It is strongly recommended that data be
filtered and normalized prior to plotting. Raw counts stored in \code{
\link[edgeR]{DGEList}} or \code{\link[DESeq2]{DESeqDataSet}} objects are
automatically extracted and transformed to the log2-CPM scale, with a
warning. Alternatively, an object of class \code{dist}.}

\item{group}{Optional character or factor vector of length equal to sample
size. Alternatively, a data frame or list of such vectors, optionally
named.}

\item{covar}{Optional continuous covariate of length equal to sample size.
Alternatively, a data frame or list of such vectors, optionally named.}

\item{dist}{Distance measure to be used. Supports all methods available in
\code{\link[stats]{dist}} and \code{\link[vegan]{vegdist}}, as well as
those implemented in the \code{bioDist} package. See Details.}

\item{p}{Power of the Minkowski distance.}

\item{top}{Optional number (if > 1) or proportion (if < 1) of top probes to
be used for distance calculations.}

\item{filter_method}{String specifying whether to apply a \code{"pairwise"}
or \code{"common"} filter if \code{top} is non-\code{NULL}. See Details.}

\item{center}{Center each probe prior to computing distances?}

\item{hclustfun}{The agglomeration method to be used for hierarchical
clustering. Supports any method available in \code{\link[stats]{hclust}}.}

\item{pal_group}{String specifying the color palette to use if \code{group}
is non-\code{NULL}, or a vector of such strings with length equal to the
number of vectors passed to \code{group}. Options include \code{"ggplot"}, 
all qualitative color schemes available in \code{\href{
https://bit.ly/2ipuEjn}{RColorBrewer}}, and the complete collection of 
\code{\href{http://bit.ly/2bxnuGB}{ggsci}} palettes. Alternatively, a 
character vector of colors with length equal to the cumulative number of 
levels in \code{group}.}

\item{pal_covar}{String specifying the color palette to use if \code{covar}
is non-\code{NULL}, or a vector of such strings with length equal to the
number of vectors passed to \code{covar}. Options include the complete 
collection of \code{\href{https://bit.ly/2n7D6tF}{viridis}} palettes, as 
well as all sequential color schemes available in \code{\href{
https://bit.ly/2ipuEjn}{RColorBrewer}}. Alternatively, a character vector 
of colors representing a smooth gradient, or a list of such vectors with 
length equal to the number of continuous variables to visualize.}

\item{pal_tiles}{String specifying the color palette to use for heatmap
tiles. Options include the complete collection of \code{\href{
https://bit.ly/2n7D6tF}{viridis}} palettes, as well all diverging color 
schemes available in \code{\href{https://bit.ly/2ipuEjn}{RColorBrewer}}. 
Alternatively, a character vector of at least two colors.}

\item{title}{Optional plot title.}
}
\description{
This function displays pairwise distances between samples as a heatmap.
}
\details{
Similarity matrices are a valuable tool for exploratory data analysis. A
hierarchical clustering dendrogram atop the figure helps identify potential
clusters and/or outliers in the data. Annotation tracks can help investigate
associations with phenotypic features.

Different distance metrics and agglomeration methods can lead to different
results. The default settings, which perform average linkage hierarchical
clustering on a Euclidean distance matrix, are mathematically straightforward
and commonly used for omic EDA. Complete linkage is also fairly common.

Other available distance measures include: \code{"maximum"}, 
\code{"manhattan"}, \code{"canberra"}, \code{"minkowski"}, \code{"cosine"},
\code{"pearson"}, \code{"kendall"}, \code{"spearman"}, \code{"bray"}, \code{
"kulczynski"}, \code{"jaccard"}, \code{"gower"}, \code{"altGower"}, \code{
"morisita"}, \code{"horn"}, \code{"mountford"}, \code{"raup"}, \code{
"binomial"}, \code{"chao"}, \code{"cao"}, \code{"mahalanobis"}, \code{"MI"},
or \code{"KLD"}. See \code{\link{dist_mat}} for more details on these methods 
and links to documentation on each.

The \code{top} argument optionally filters data using either probewise
variance (if \code{filter_method = "common"}) or the leading fold change
method of Smyth et al. (if \code{filter_method = "pairwise"}). See \code{
\link{plot_mds}} for more details.
}
\examples{
mat <- matrix(rnorm(5000), nrow = 1000, ncol = 5)
plot_similarity(mat, title = "Nothin' Doin'")

library(DESeq2)
dds <- makeExampleDESeqDataSet()
dds <- rlog(dds)
plot_similarity(dds, group = colData(dds)$condition,
                title = "Somethin' Cookin'")

}
