% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Create volcano plot of \eqn{log_2} fold changes against \eqn{-log_10 p}-values}
\usage{
plot_volcano(dat, fdr = 0.05, ptsize = 0.25, main = NULL,
  legend = "outside")
}
\arguments{
\item{dat}{Data frame or matrix representing the results of a test for 
differential expression or methylation, such as the output of a call to 
limma::topTable, edgeR::topTags, or DESeq2::results. Alternatively, any object 
with columns for p-values, log fold changes, and FDR.}

\item{fdr}{Threshold for declaring a probe differentially expressed/methylated.}

\item{ptsize}{Size of data points in the plot.}

\item{main}{Optional plot title.}

\item{legend}{Legend position. Must be one of \code{"outside", 
"bottomleft", "bottomright", "topleft",} or \code{"topright"}.}
}
\description{
Create volcano plot of \eqn{log_2} fold changes against \eqn{-log_10 p}-values
}
\details{
This function displays the results of a differential expression or methylation 
test as a volcano plot, with colors distinguishing probes that meet the 
user-defined FDR threshold. These figures can be used to evaluate the symmetry,
magnitude, and significance of fold changes for a given experiment.
}
\examples{
library(dplyr)
df <- data_frame(logFC   = c(rnorm(50, 0, 10), rnorm(4950)),
                 p.value = pnorm(-abs(logFC)),
                 FDR     = p.adjust(p.value, method = 'fdr'))
plot_volcano(df)

library(limma)
DE_genes <- cbind(matrix(rnorm(250, 5, 1), nrow = 50, ncol = 5),
                  matrix(rnorm(250), nrow = 50, ncol = 5))
mat <- rbind(DE_genes, matrix(rnorm(45500), nrow = 4550, ncol = 10))
treat <- c(rep("A", 5), rep("B", 5))
des <- model.matrix(~ treat)
fit <- eBayes(lmFit(mat, des))
top <- topTable(fit, number = Inf)
plot_volcano(top)

}

