% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Volcano Plot}
\usage{
plot_volcano(dat, fdr = 0.05, ptsize = 0.25, main = NULL,
  legend = "outside", hover = FALSE)
}
\arguments{
\item{dat}{Data frame or matrix representing the results of a test for
differential expression or methylation, such as the output of a call to
\code{\link[limma]{topTable}}, \code{\link[edgeR]{topTags}}, or
\code{\link[DESeq2]{results}}. Alternatively, any object with columns for
\emph{p}-values, log fold changes, and FDR. \code{NA} values are silently
removed.}

\item{fdr}{Threshold for declaring a probe differentially expressed/methylated.}

\item{ptsize}{Size of data points in the plot.}

\item{main}{Optional plot title.}

\item{legend}{Legend position. Must be one of \code{"outside",
"bottomleft", "bottomright", "topleft",} or \code{"topright"}.}

\item{hover}{Show probe name by hovering mouse over data point? If \code{TRUE},
the plot is rendered in HTML and will either open in your browser's graphic
display or appear in the RStudio viewer. Probe names are extracted from
\code{dat}.}
}
\description{
This function plots the log2 fold changes against the -log10 \emph{p}-values
for a given test of differential expression/methylation.
}
\details{
Volcano plots visualize the relationship between each probe's log2 fold change and
-log10 \emph{p}-value for a given test of differential expression/methylation.
Points are colored to distinguish between those that do and do not meet a
user-defined FDR threshold. These figures help to evaluate the symmetry,
magnitude, and significance of effects for a given experiment.
}
\examples{
library(limma)
DE_genes <- cbind(matrix(rnorm(50 * 5, mean = 5), nrow = 50, ncol = 5),
                  matrix(rnorm(50 * 5), nrow = 50, ncol = 5))
eset <- rbind(DE_genes, matrix(rnorm(4950 * 10), nrow = 4950, ncol = 10))
treat <- rep(c("A", "B"), each = 5)
des <- model.matrix(~ treat)
fit <- eBayes(lmFit(eset, des))
top <- topTable(fit, number = Inf)
plot_volcano(top)

}
