% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stringify}
\alias{stringify}
\title{Format String}
\usage{
stringify(x)
}
\arguments{
\item{x}{Character vector.}
}
\description{
This utility function formats a character vector into a string representing
a well formed English list, with quotations around each element and an "and"
between the penultimate and ultimate elements.
}
